% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anes2016_sample_corpus}
\alias{anes2016_sample_corpus}
\title{ANES 2016 sample corpus}
\format{
A quanteda corpus with 3600 documents and 4 docvars:
\describe{
\item{respondent_id}{respondent id (do not correspond to the original ids)}
\item{ideology}{ideology of respondent, Liberal, Moderate, Conservative}
\item{yob}{year of birth}
\item{generation}{generation defined by yob, see pre-processing file}
...
}
}
\source{
\url{https://electionstudies.org/data-center/2016-time-series-study/}
}
\usage{
anes2016_sample_corpus
}
\description{
A (quanteda) corpus containing a sample of the ANES 2016 open-ended
responses to the question:
"what is the most important problem facing the country?"
300 responses for each ideology - generation pair is sampled.
For full data and pre-processing file, see:
https://www.dropbox.com/sh/jsyrag7opfo7l7i/AAB1z7tumLuKihGu2-FDmhmKa?dl=0
}
\keyword{datasets}
