% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize-col.r
\name{binarize_col}
\alias{binarize_col}
\title{Convert single variable columns to multiple binary columns}
\usage{
binarize_col(data, col = 1:ncol(data), sep = ".", na.value = 0)
}
\arguments{
\item{data}{Data frame.}

\item{col}{Character or numeric. The names or indices of the columns of
\code{data} to be converted.}

\item{sep}{Character. The string to connect each variable to its value to
create names for the binary columns.}

\item{na.value}{Numeric. Value to assign to NA entries in value-specific
columns. (NA entries are converted to 1 in NA columns.) Defaults to 0.}
}
\value{
A data frame having the same number of rows as \code{data}.
}
\description{
Convert single variable columns to multiple binary columns
}

