% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_nns.R
\name{bootstrap_nns}
\alias{bootstrap_nns}
\title{Bootstrap nearest neighbors}
\usage{
bootstrap_nns(
  context = NULL,
  pre_trained = NULL,
  transform = TRUE,
  transform_matrix = NULL,
  candidates = NULL,
  bootstrap = TRUE,
  num_bootstraps = 20,
  N = 50,
  norm = "l2"
)
}
\arguments{
\item{context}{vector of texts - `context`` variable in get_context output}

\item{pre_trained}{a V x D matrix of numeric values - pretrained embeddings with V = size of vocabulary and D = embedding dimensions}

\item{transform}{logical - if TRUE (default) apply the a la carte transformation, if FALSE ouput untransformed averaged embedding}

\item{transform_matrix}{a D x D transformation matrix}

\item{candidates}{character vector defining the candidates for nearest neighbors - e.g. output from \code{get_local_vocab}}

\item{bootstrap}{logical - if TRUE, bootstrap cosine similarity - required to get standard errors cosine similarity of nearest neighbors}

\item{num_bootstraps}{numeric - number of bootstraps to use}

\item{N}{number of nearest neighbors to return}

\item{norm}{character = c("l2", "none") - set to 'l2' for cosine similarity and to 'none' for inner product (see ?sim2 in text2vec)}
}
\value{
data.frame with nearest neighbor tokens, similarity (numeric value) and std.error (given bootstrap)
}
\description{
Bootstrap nearest neighbors
}
\examples{

immig_corpus <- corpus_context(x = cr_sample_corpus,
pattern = "immigration", window = 6L, verbose = TRUE)

set.seed(42L)
bootstrap_nns(context = immig_corpus,
              pre_trained = glove_subset,
              transform = TRUE, transform_matrix = khodakA,
              bootstrap = TRUE, num_bootstraps = 20,
              N = 20, norm = "l2")

}
