% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_nns.R
\name{bootstrap_similarity}
\alias{bootstrap_similarity}
\title{Boostrap similarity vector}
\usage{
bootstrap_similarity(
  target_embeddings = NULL,
  pre_trained = NULL,
  candidates = NULL,
  norm = NULL
)
}
\arguments{
\item{target_embeddings}{the target embeddings (embeddings of context)}

\item{pre_trained}{a V x D matrix of numeric values - pretrained embeddings with V = size of vocabulary and D = embedding dimensions}

\item{candidates}{character vector defining the candidates for nearest neighbors - e.g. output from \code{get_local_vocab}}

\item{norm}{character = c("l2", "none") - set to 'l2' for cosine similarity and to 'none' for inner product (see ?sim2 in text2vec)}
}
\value{
vector(s) of cosine similarities between alc embedding and nearest neighbor candidates
}
\description{
Boostrap similarity vector
}
