% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-builder.R
\name{build_conText}
\alias{build_conText}
\title{build a \linkS4class{conText} object}
\usage{
build_conText(
  Class = "conText",
  x_conText,
  normed_cofficients = data.frame(),
  features = character(),
  Dimnames = list()
)
}
\arguments{
\item{Class}{defines the class of this object (fixed)}

\item{normed_cofficients}{a data.frame withe the normed coefficients and
other statistics}

\item{features}{features used in computing the embeddings}

\item{Dimnames}{row (features) and columns (NULL) names}

\item{x}{a \link{dgCMatrix class} matrix}
}
\description{
build a \linkS4class{conText} object
}
