% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-builder.R
\name{build_dem}
\alias{build_dem}
\title{build a \code{dem-class} object}
\usage{
build_dem(
  Class = "em",
  x_dem,
  docvars = data.frame(),
  features = character(),
  Dimnames = list()
)
}
\arguments{
\item{Class}{defines tha class of this object (fixed)}

\item{x_dem}{a \verb{dgCMatrix class} matrix}

\item{docvars}{document covariates, inherited from dfm and corpus,
subset to embeddable documents}

\item{features}{features used in computing the embeddings}

\item{Dimnames}{row (documents) and columns (NULL) names}
}
\description{
build a \code{dem-class} object
}
