% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-builder.R
\name{build_fem}
\alias{build_fem}
\title{build a \linkS4class{fem} object}
\usage{
build_fem(Class = "fem", x_fem, features = character(), Dimnames = list())
}
\arguments{
\item{Class}{defines the class of this object (fixed)}

\item{features}{features used in computing the embeddings}

\item{Dimnames}{row (features) and columns (NULL) names}

\item{x}{a \link{dgCMatrix class} matrix}
}
\description{
build a \linkS4class{fem} object
}
