% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drat.R
\name{build_local_drat}
\alias{build_local_drat}
\title{Build local drat repository}
\usage{
build_local_drat(sources, path, force = FALSE, quiet = TRUE)
}
\arguments{
\item{sources}{A \code{\link{package_sources}} object}

\item{path}{Path to build the drat repository at}

\item{force}{Re-fetch all packages, even if they still look fresh.}

\item{quiet}{Quieten the download progress (which can be quite messy)}
}
\description{
Build a local ad-hoc drat from a set of package sources.  This
downloads the zipball from github, bitbucket, or locates it
locally, builds a \emph{source} package and adds it to an ad-hoc
drat repository within \code{path}.  This is designed to be used
within \code{install_packages} and \code{context_save}, but can be
called directly.
}

