% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_bulk.R
\name{bulk_prepare_expression}
\alias{bulk_prepare_expression}
\title{Prepare many expressions}
\usage{
bulk_prepare_expression(X, FUN, DOTS, do_call, use_names, envir, db)
}
\arguments{
\item{X}{Something to iterate over; a vector, list or data.frame
(in the case of a data.frame, iteration will be row-by-row)}

\item{FUN}{A function to apply to each element (or row) of
\code{X}}

\item{DOTS}{Additional arguments to apply with each elements of
\code{X}}

\item{do_call}{Treat each element of \code{X} as a \code{do.call} call}

\item{use_names}{When preparing a data.frame, retain column names
as argument names when using \code{do_call}.  If \code{FALSE}
then positional matching will be used.}

\item{envir}{An environment to find variables local to the expression}

\item{db}{A database to store locals}
}
\description{
Prepare many expressions
}
