% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_bulk.R
\name{bulk_task_save}
\alias{bulk_task_save}
\title{Save bulk tasks}
\usage{
bulk_task_save(
  X,
  FUN,
  context,
  DOTS = NULL,
  do_call = FALSE,
  use_names = TRUE,
  envir = parent.frame(),
  depends_on = NULL
)
}
\arguments{
\item{X}{Something to iterate over; a vector, list or data.frame
(in the case of a data.frame, iteration will be row-by-row)}

\item{FUN}{A function to apply to each element (or row) of
\code{X}}

\item{context}{A context}

\item{DOTS}{Additional arguments to apply with each elements of
\code{X}}

\item{do_call}{Treat each element of \code{X} as a \code{do.call} call}

\item{use_names}{When preparing a data.frame, retain column names
as argument names when using \code{do_call}.  If \code{FALSE}
then positional matching will be used.}

\item{envir}{An environment to find variables local to the expression}

\item{depends_on}{Optional task ids that this task
depends on. To have all tasks depend on the same id(s)
provide a vector. TO provide different dependencies for each
task provide a list of lists.
For example list(list("abcde", "12345"), list(), list("12345"))}
}
\description{
Save bulk tasks
}
