% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{conText-class}
\alias{conText-class}
\title{Virtual class "conText" for a conText regression output}
\description{
The \code{conText-class} is \verb{dgCMatrix class} matrix corresponding to
the beta coefficients (embeddings) with additional slots:
}
\section{Slots}{

\describe{
\item{\code{normed_coefficients}}{\code{normed_betas} a data.frame containing the following variables:
\describe{
\item{\code{Coefficient}}{(character) non-intercept coefficient names}
\item{\code{Normed_Estimate}}{(numeric) norm of non-intercept beta coefficients}
\item{\code{Std.Error}}{(numeric) std errors (given boostrap)}
\item{\code{Empirical_Pvalue}}{(numeric) empirical pvalue (given permute)}
}}

\item{\code{features}}{features used in computing the document embeddings}
}}

\seealso{
\code{conText}
}
\keyword{conText}
\keyword{internal}
