% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conText-package.R
\docType{package}
\name{conText-package}
\alias{conText-package}
\title{conText: 'a la Carte' on Text (ConText) Embedding Regression}
\description{
A fast, flexible and transparent framework to estimate context-specific word and short document embeddings using the 'a la carte' embeddings approach developed by Khodak et al. (2018) \doi{10.48550/arXiv.1805.05388} and evaluate hypotheses about covariate effects on embeddings using the regression framework developed by Rodriguez et al. (2021)\doi{10.1017/S0003055422001228}. New version of the package applies a new estimator to measure the distance between word embeddings as described in Green et al. (2025) \doi{10.1017/pan.2024.22}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/prodriguezsosa/conText}
  \item Report bugs at \url{https://github.com/prodriguezsosa/ConText/issues}
}

}
\author{
\strong{Maintainer}: Sofia Avila \email{sofiaavila@princeton.edu} (\href{https://orcid.org/0009-0007-8746-6688}{ORCID})

Authors:
\itemize{
  \item Pedro L. Rodriguez \email{pedro.rodriguezsosa@gmail.com} (\href{https://orcid.org/0000-0003-4762-4550}{ORCID}) [copyright holder]
  \item Arthur Spirling \email{as9934@nyu.edu} (\href{https://orcid.org/0000-0001-9959-1805}{ORCID})
  \item Brandon Stewart \email{bms4@princeton.edu} (\href{https://orcid.org/0000-0002-7657-3089}{ORCID})
}

Other contributors:
\itemize{
  \item Christopher Barrie \email{christopher.barrie@ed.ac.uk} (\href{https://orcid.org/0000-0002-9156-990X}{ORCID}) [contributor]
}

}
\keyword{internal}
