% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context_list}
\alias{context_info}
\alias{context_list}
\title{List save contexts}
\usage{
context_list(db, named = FALSE, error = TRUE)

context_info(db, error = TRUE)
}
\arguments{
\item{db}{Something for which a context database can be created;
this can the the path to the context, a \code{context_root}
object, or a \code{context} object.}

\item{named}{Logical, indicating if the context name should be used
to name the output vector.}

\item{error}{Throw an error if the context database cannot be
connected constructed (e.g., if the path given does not exist).}
}
\description{
List saved contexts
}
\author{
Rich FitzJohn
}

