% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context_load}
\alias{context_load}
\title{Load a context}
\usage{
context_load(ctx, envir = .GlobalEnv, refresh = FALSE)
}
\arguments{
\item{ctx}{A context object, as read by \code{\link{context_read}}}

\item{envir}{The environment to source files into}
}
\description{
Load a context
}

