% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{context_log_start}
\alias{context_log_start}
\alias{context_log_stop}
\alias{parse_context_log}
\title{Start and stop log}
\usage{
context_log_start()

context_log_stop()

parse_context_log(x)
}
\arguments{
\item{x}{Vector of log output}
}
\value{
\code{context_log_start} invisibly returns a logical
  indicating if logging was previously enabled.  This allows
  patterns like:
\preformatted{if (!context::context_log_start()) {
  context::context_log_stop()
}
}
to have a scoped log (i.e., log for the duration of a function).
}
\description{
Start and stop the context log.  Soon this might swap out for
\code{loggr}, but for now this should do.  When active, some
actions will print diagnostic information to the message stream.
This is used in particular by the program installed by
\code{\link{install_context}}.
}
\details{
The interface here will change by adding arguments.  Future versions
may support logging to a file.
}

