% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context_read}
\alias{context_read}
\title{Read a context}
\usage{
context_read(identifier, root, db = NULL)
}
\arguments{
\item{identifier}{Either the id or name of a context (see
\code{\link{context_list}})}

\item{root}{Something interpretable as the context root; either}

\item{db}{Optionally, a database (if known already)}
}
\description{
Read a context
}
