% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_root.R
\name{context_root_get}
\alias{context_root_get}
\title{Find context root}
\usage{
context_root_get(root, db = NULL)
}
\arguments{
\item{root}{An object; either a character string (interpreted as a
path), a \code{context_root} object (such as returned by this
function) or a list/environment object with a \code{root}
element that is a \code{context_root} object.}

\item{db}{Optionally, a copy of the storr database (if already
opened).  Do not specify this unless you \emph{definitely} have
the correct database in hand.}
}
\description{
Find the context root.  Designed for internal use
}
