% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context_save}
\alias{context_save}
\title{Save a context}
\usage{
context_save(path, packages = NULL, sources = NULL, auto = FALSE,
  package_sources = NULL, envir = NULL, storage_type = NULL,
  storage_args = NULL, name = NULL)
}
\arguments{
\item{path}{Path to save the context in}

\item{packages}{Optional character vector of packages to save into
the context.  Alternatively, can be a list with elements
\code{loaded} and \code{attached} if you want to ensure some
packages are loaded but not attached.}

\item{sources}{Character vector of source files to read in.  These
should define functions and (perhaps) other "global" objects,
but should not do any serious computation.}

\item{auto}{Attempt to create the context automatically.  In this
case, do not specify either \code{packages} or \code{sources}.}

\item{package_sources}{Optional information about where to find
non-CRAN packages.}

\item{envir}{The current environment.  This is used to copy
\emph{local} enviroments around, as these are needed even in the
case of non-automatic contexts.  For \code{context_load} this is
the environment into which the global environment is copied.
Specify a non-global environment here to avoid clobbering the
workspace, but at the risk that some environments may not
restore exactly as desired.}

\item{storage_type}{Character vector indicating the storage type
to use.  Options are \code{"rds"} (the default) and
\code{"environment"} (for testing and local use).}

\item{storage_args}{Arguments passed through to the storage driver}

\item{name}{An optional name for the context.  This will be
printed with the context in some situations (such as
\code{\link{context_info}})}
}
\description{
Save a context
}

