% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_context.R
\name{corpus_context}
\alias{corpus_context}
\title{Get contexts sorrounding a user defined keyword along with typicality scores}
\usage{
corpus_context(
  x,
  pattern,
  window = 6L,
  valuetype = c("glob", "regex", "fixed"),
  case_insensitive = TRUE,
  hard_cut = FALSE,
  exclude_pattern = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{a \link{corpus} object}

\item{pattern}{a character vector, list of character vectors, \link[quanteda]{dictionary},
or collocations object.  See \link[quanteda]{pattern} for details.}

\item{window}{the number of context words to be displayed around the keyword}

\item{valuetype}{the type of pattern matching: \code{"glob"} for "glob"-style
wildcard expressions; \code{"regex"} for regular expressions; or \code{"fixed"} for
exact matching. See \link[quanteda]{valuetype} for details.}

\item{case_insensitive}{logical; if \code{TRUE}, ignore case when matching a
\code{pattern} or \link[quanteda]{dictionary} values}

\item{hard_cut}{logical - if TRUE then the text must have window x 2 tokens,
if FALSE it can have window x 2 or fewer (e.g. if a doc begins with a target word,
then text will have window tokens rather than window x 2)}

\item{exclude_pattern}{(logical) if TRUE, pattern/keyword is excluded from the text documents}

\item{verbose}{(logical) if TRUE, report the total number of target instances found.}
}
\value{
a data.frame with with the same columns as quanteda::kwic.
If typicality_score = TRUE, then an additional (numeric) column with
typicality scores is added and data.frame is sorted from
highest to lowest typicality score.
}
\description{
A wrapper for quanteda::kwic() with two added functionalities:
i. subsets documents to where keyword is present before tokenizing thereby speeding up processing
ii. compute typicality scores (average cosine similarity between a context and all other contexts)
Contexts with higher typicality scores are more representative.
}
\examples{

immig_corpus <- corpus_context(x = cr_sample_corpus,
pattern = "immigration", window = 6L, verbose = TRUE)

}
\keyword{corpus_context}
