% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_sim.R
\name{cos_sim}
\alias{cos_sim}
\title{Given a set of embeddings and a set of features, calculate cosine similarities.}
\usage{
cos_sim(x, pre_trained, features = NULL)
}
\arguments{
\item{x}{a \linkS4class{dem} or \linkS4class{fem} object}

\item{pre_trained}{a F x D matrix of numeric values corresponding to pretrained embeddings
F = number of features and D = embedding dimensions.
rownames(pre_trained) = set of features for which there is a pre-trained embedding}

\item{features}{(character) features of interest}
}
\value{
a \code{data.frame} with following columns:
\item{\code{target}}{ (character) vector with the rownames of the dfm,
either defining the groups or the target terms}.
\item{\code{feature}}{(character) vector of feature terms, one
instance for each target.}
\item{\code{value}}{(numeric) cosine similarity between target
and feature.}
}
\description{
Given a set of embeddings and a set of features, calculate cosine similarities.
}
\examples{

library(quanteda)

# build corpus of contexts around immigration
immig_corpus <- corpus_context(x = cr_sample_corpus,
pattern = "immigration", window = 6L, verbose = TRUE)

# tokenize text
immig_toks <- tokens(immig_corpus)

# construct document-feature-matrix
immig_dfm <- dfm(immig_toks)

# construct document-embedding-matrix
immig_dem <- dem(immig_dfm, pre_trained = glove_subset,
transform = TRUE, transform_matrix = khodakA, verbose = FALSE)

# group document-embedding-matrix
immig_dem_party <- dem_group(immig_dem,
groups = immig_dem@docvars$party)

# find nearest neighbors
cos_sim(x = immig_dem_party,
pre_trained = glove_subset, features = c('reform', 'enforce'))
}
\keyword{cos_sim}
