% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cr_glove_subset}
\alias{cr_glove_subset}
\title{GloVe subset}
\format{
A matrix with 500 rows and 300 columns:
\describe{
\item{row}{each row corresponds to a word}
\item{column}{each column corresponds to a dimension in the embedding space}
...
}
}
\source{
\url{https://www.dropbox.com/s/p84wzv8bdmziog8/cr_glove.R?dl=0}
}
\usage{
cr_glove_subset
}
\description{
A subset of a GloVe embeddings model trained on the top 5000 features
in the Congressional Record Record corpus covering the 111th - 114th Congresses,
and limited to speeches by Democrat and Republican representatives.
}
\keyword{datasets}
