% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cr_transform}
\alias{cr_transform}
\title{Transformation matrix}
\format{
A 300 by 300 matrix.
}
\source{
\url{https://www.dropbox.com/s/p84wzv8bdmziog8/cr_glove.R?dl=0}
}
\usage{
cr_transform
}
\description{
A square matrix corresponding to the transformation matrix computed
using the cr_glove_subset embeddings and corresponding corpus.
}
\keyword{datasets}
