% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{create_context_manager}
\alias{create_context_manager}
\title{Context manager}
\usage{
create_context_manager(
  on_enter = function() {
 },
  on_exit = function() {
 },
  args = NULL
)
}
\arguments{
\item{on_enter}{function to run before entering context}

\item{on_exit}{function to run after taking actions}

\item{args}{Optional list of arguments supplied to `on_enter` function.}
}
\value{
S3 class ContextManager
}
\description{
Context manager
}
