% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_install.R
\name{cross_install_packages}
\alias{cross_install_bootstrap}
\alias{cross_install_context}
\alias{cross_install_packages}
\title{Cross install packages}
\usage{
cross_install_packages(lib, platform, r_version, repos, packages)

cross_install_bootstrap(lib, platform, r_version, root = NULL)

cross_install_context(lib, platform, r_version, context)
}
\arguments{
\item{lib}{A library to install into.  \emph{Must} be specified
(you do not want to install these into your usual library!)}

\item{platform}{The platform to target.  Must be one of "windows",
"macosx" or "macosx/mavericks"}

\item{r_version}{The R version to target, of form "X.Y" (ignores
patch version).}

\item{repos}{Repositories to use.  This must be provided at present.}

\item{packages}{Vector of packages to install}

\item{root}{Context root}

\item{context}{A context handle}
}
\description{
Support for cross installing packages (e.g., installing windows
packages into a library while running on Linux).  This is
experimental and there are rather a lot of corner cases that
aren't covered.
}

