% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.R
\name{dem}
\alias{dem}
\title{Create an document-embedding matrix}
\usage{
dem(x, pre_trained, transform = TRUE, transform_matrix, verbose = TRUE)
}
\arguments{
\item{x}{a \code{dfm-class} object}

\item{pre_trained}{a F x D matrix of numeric values corresponding to pretrained embeddings
F = number of features and D = embedding dimensions.
rownames(pre_trained) = set of features for which there is a pre-trained embedding}

\item{transform}{(logical) if TRUE (default) apply the \verb{a la carte} transformation,
if FALSE ouput untransformed averaged embedding}

\item{transform_matrix}{a D x D matrix of numeric values corresponding
to the `a la carte`` transformation matrix}

\item{verbose}{logical - report the documents/features that had
no overlap with the provided pre-trained embeddings}
}
\value{
a \code{dem-class} object
}
\description{
Given a document-feature matrix for each document,
multiply its feature counts (columns) with their
corresponding pre-trained word embeddings and average
(usually referred to as averaged or additive document embeddings).
If specified and a transformation matrix is provided,
multiply the document embeddings by the transformation matrix
to obtain the corresponding \verb{a la carte} document embeddings.
(see eq 2: https://arxiv.org/pdf/1805.05388.pdf)
}
\examples{

library(quanteda)

# tokenize corpus
toks <- tokens(cr_sample_corpus)

# build a tokenized corpus of contexts sorrounding a target term
immig_toks <- tokens_context(x = toks, pattern = "immigr*", window = 6L)

# construct document-feature-matrix
immig_dfm <- dfm(immig_toks)

# construct document-embedding-matrix
immig_dem <- dem(immig_dfm, pre_trained = cr_glove_subset,
transform = TRUE, transform_matrix = cr_transform, verbose = FALSE)
}
\keyword{dem}
