% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_group.R
\name{dem_group}
\alias{dem_group}
\title{Average document-embeddings in a dem by a grouping variable}
\usage{
dem_group(x, groups = NULL)
}
\arguments{
\item{x}{a \code{dem}}

\item{groups}{a character or factor variable equal in length to the number of documents}
}
\value{
a \code{dem-class} object
}
\description{
Average embeddings in a dem by a grouping variable, by averaging over columns within groups
and creating new "documents" with the group labels.
Similar in essence to \code{dfm_group}.
}
\examples{

library(quanteda)

# tokenize text
cr_toks <- tokens(cr_sample_corpus)

# construct document-feature-matrix
cr_dfm <- dfm(cr_toks)

# construct document-embedding-matrix
cr_dem <- dem(cr_dfm, pre_trained = cr_glove_subset,
transform = TRUE, transform_matrix = cr_transform, verbose = FALSE)

# group document-embedding-matrix
cr_dem_party <- dem_group(cr_dem,
groups = cr_dem@docvars$party)

}
\keyword{dem_group}
