% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ejem}
\alias{ejem}
\title{Drive concepts for motorcars.}
\format{A many-valued context, stored as a data frame.
\describe{
  \item{De}{drive efficiency empty}
  \item{Dl}{drive efficiency loaded}
  \item{R}{road holding/handling properties}
  \item{S}{self-steering effect}
  \item{E}{economy of space}
  \item{C}{cost of construction}
  \item{M}{maintainability}
}}
\source{
Ganter, B. and Stumme, G. and Wille, R. (2005) Formal Concept
  Analysis: Foundations and Applications
  \url{http://www.springer.com/us/book/9783540278917}
}
\usage{
ejem
}
\description{
This many-valued context is adapted from an example of conceptual scaling.
}
\examples{
# scale all variables using nominal standard scales
data(ejem)
variables <- colnames(ejem)
ejem_scale <- scale_by_variables(ejem,
                                 vars = variables,
                                 scales = as.list(rep('nominal',
                                                      length(variables))))
rownames(ejem_scale) <- rownames(ejem)
print(ejem_scale)
}
\keyword{datasets}

