% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{fem-class}
\alias{fem-class}
\title{Virtual class "fem" for a feature-embedding matrix}
\description{
The \link{fem class} is \link{dgCMatrix class} matrix with additional slots:
}
\section{Slots}{

\describe{
\item{\code{features}}{features used in computing the document embeddings}
}}

\seealso{
\link{fem}
}
\keyword{fem}
\keyword{internal}
