% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fem.R
\name{fem}
\alias{fem}
\title{Create an feature-embedding matrix}
\usage{
fem(x, pre_trained, transform = TRUE, transform_matrix, verbose = TRUE)
}
\arguments{
\item{x}{a \linkS4class{fcm} object}

\item{pre_trained}{a F x D matrix of numeric values corresponding to pretrained embeddings
F = number of features and D = embedding dimensions.
rownames(pre_trained) = set of features for which there is a pre-trained embedding}

\item{transform}{(logical) if TRUE (default) apply the \link{a la carte} transformation, if FALSE ouput untransformed averaged embedding}

\item{transform_matrix}{a D x D matrix of numeric values corresponding to the `a la carte`` transformation matrix}

\item{verbose}{logical - report the documents/features that had no overlap with the provided pre-trained embeddings}
}
\value{
a \linkS4class{fem} object
}
\description{
Given a featureco-occurrence matrix for each feature,
multiply its feature counts (columns) with their
corresponding pre-trained embeddings and average
(usually referred to as averaged or additive embeddings).
If specified and a transformation matrix is provided,
multiply the feature embeddings by the transformation matrix
to obtain the corresponding \verb{a la carte} embeddings.
(see eq 2: https://arxiv.org/pdf/1805.05388.pdf)
}
\examples{
## embed all features
corpus_fcm <- create_fcm(cr_corpus_sample, features_remove = quanteda::stopwords('en'))
corpus_fem <- fem(corpus_fcm, pre_trained = glove_subset, transform = TRUE, transform_matrix = khodakA, verbose = TRUE)

## embed specific target features
corpus_fcm <- create_fcm(cr_corpus_sample, target = c('immigration','welfare'), features_remove = quanteda::stopwords('en'))
corpus_fem <- fem(corpus_fcm, pre_trained = glove_subset, transform = TRUE, transform_matrix = khodakA, verbose = TRUE)
}
\keyword{fem}
