% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seq_cos_sim.R
\name{find_cos_sim}
\alias{find_cos_sim}
\title{Find cosine similarities between target and candidate words}
\usage{
find_cos_sim(target_embedding, pre_trained, candidates, norm = "l2")
}
\arguments{
\item{target_embedding}{matrix of numeric values}

\item{pre_trained}{matrix of numeric values - pretrained embeddings}

\item{candidates}{character vector defining vocabulary to subset comparison to}

\item{norm}{character = c("l2", "none") - how to scale input matrices. If they are already scaled - use "none" (see ?sim2)}
}
\value{
a vector of cosine similarities of length candidates
}
\description{
Find cosine similarities between target and candidate words
}
