% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nns.R
\name{find_nns}
\alias{find_nns}
\title{Return nearest neighbors based on cosine similarity}
\usage{
find_nns(target_embedding, pre_trained, N = 5, candidates = NULL, norm = "l2")
}
\arguments{
\item{target_embedding}{matrix of numeric values}

\item{pre_trained}{matrix of numeric values - pretrained embeddings}

\item{N}{number of nearest neighbors to return}

\item{candidates}{character vector defining vocabulary to subset conmparison to}

\item{norm}{character = c("l2", "none") - how to scale input matrices. If they are already scaled - use "none" (see ?sim2)}
}
\value{
a character vector of nearest neighbors to target
}
\description{
Return nearest neighbors based on cosine similarity
}
\examples{
find_nns(target_embedding = glove_subset['immigration',],
         pre_trained = glove_subset, N = 5,
         candidates = NULL, norm = "l2")
}
