% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_vocab.R
\name{get_local_vocab}
\alias{get_local_vocab}
\title{Get local vocab}
\usage{
get_local_vocab(context, pre_trained)
}
\arguments{
\item{context}{the text to be used to compute localized embeddings}

\item{pre_trained}{the pre-trained embeddings to be used to compute localized embeddings}
}
\value{
chr vector - vector of words in vocabulary
}
\description{
Local vocab consists of the intersect between the set of pre-trained embeddings
and the local contexts.
}
\examples{
library(conText)
library(dplyr)

# load data
corpus <- sample_corpus
pre_trained <- sample_glove
transform_matrix <- khodakA

# find contexts of immigration
context_immigration <- get_context(x = corpus$speech, target = 'immigration',
                        window = 6, valuetype = "fixed", case_insensitive = TRUE,
                        hard_cut = FALSE, verbose = FALSE)

# find local vocab (use it to define the candidate of nearest neighbors)
local_vocab <- get_local_vocab(context_immigration$context, pre_trained)
}
