% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nns.R
\name{get_nns}
\alias{get_nns}
\title{Given a corpus and a set of candidate neighbors, find the top N nearest
neighbors.}
\usage{
get_nns(
  x,
  N = 10,
  groups = NULL,
  candidates = character(0),
  pre_trained,
  transform = TRUE,
  transform_matrix,
  bootstrap = TRUE,
  num_bootstraps = 10,
  as_list = TRUE
)
}
\arguments{
\item{x}{a (quanteda) tokens object}

\item{N}{(numeric) number of nearest neighbors to return}

\item{groups}{grouping variable equal in length to the number of documents}

\item{candidates}{(character) vector of features to consider as candidates to be nearest neighbor
You may for example want to only consider features that meet a certian count threshold
or exclude stop words etc. To do so you can simply identify the set of features you
want to consider and supply these as a character vector in the \code{candidates} argument.}

\item{pre_trained}{a F x D matrix of numeric values corresponding to pretrained embeddings
F = number of features and D = embedding dimensions.
rownames(pre_trained) = set of features for which there is a pre-trained embedding}

\item{transform}{(logical) if TRUE (default) apply the \verb{a la carte} transformation,
if FALSE ouput untransformed averaged embedding}

\item{transform_matrix}{a D x D matrix of numeric values corresponding
to the `a la carte`` transformation matrix}

\item{bootstrap}{(logical) if TRUE, bootstrap nns - sample from corpus with replacement;
if groups defined, sampling is automatically stratified; top nns are those with
the highest average over all bootstrap samples.}

\item{num_bootstraps}{(integer) number of bootstraps to use}

\item{as_list}{(logical) if FALSE all results are combined into a single data.frame
If TRUE, a list of data.frames is returned with one data.frame per target.}
}
\value{
a \code{data.frame} or list of data.frames (one for each target)
with the following columns:
\item{\code{target}}{ (character) vector with the rownames of the dfm,
either defining the groups or the target terms}.
\item{\code{feature}}{(character) vector of features from the candidate set,
one instance for each target.}
\item{\code{value}}{(numeric) cosine similarity between target
and candidate (mean of boostraps if boostrap = TRUE).}
\item{\code{std.error}}{(numeric) sd of bootstrapped cosine similarities
if bootstrap = TRUE, if FALSE, column is dropped.}
}
\description{
Given a corpus and a set of candidate neighbors, find the top N nearest
neighbors.
}
\examples{

library(quanteda)

# tokenize corpus
toks <- tokens(cr_sample_corpus)

# build a tokenized corpus of contexts sorrounding a target term
immig_toks <- tokens_context(x = toks, pattern = "immigr*", window = 6L)

# we limit candidates to features in our corpus
feats <- featnames(dfm(immig_toks))

# compare nearest neighbors between groups
set.seed(2021L)
immig_party_nns <- get_nns(x = immig_toks, N = 10,
                           groups = docvars(immig_toks, 'party'),
                           candidates = feats,
                           pre_trained = cr_glove_subset,
                           transform = TRUE,
                           transform_matrix = cr_transform,
                           bootstrap = TRUE,
                           num_bootstraps = 10,
                           as_list = TRUE)

# nearest neighbors of "immigration" for Republican party
immig_party_nns[["R"]]
}
\keyword{get_nns}
