% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glove_subset}
\alias{glove_subset}
\title{GloVe subset}
\format{
A matrix with 393 rows and 300 columns:
\describe{
\item{row}{each row corresponds to a word}
\item{column}{each column corresponds to a dimension in the embedding space}
...
}
}
\source{
\url{https://nlp.stanford.edu/projects/glove/}
}
\usage{
glove_subset
}
\description{
A subset of Stanford GloVe embeddings containing the top terms in the sample corpora.
}
\keyword{datasets}
