% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{\%as\%}
\alias{\%as\%}
\title{Context manager's `as` function}
\usage{
context \%as\% variables
}
\arguments{
\item{context}{Context manager constructor.}

\item{variable}{Variable name to use within the context.}
}
\value{
ContextAs S3 object  with elements `context` and `variable`.
}
\description{
Context manager's `as` function
}
