% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{install_context}
\alias{install_context}
\title{Install script}
\usage{
install_context(path)
}
\arguments{
\item{path}{Path to install the script, ideally on your \code{$PATH}.
This directory must exist.}
}
\description{
Install a script that runs the context.  This won't work on
windows, but I'll get that sorted soon.
}
\details{
This script requires that context (and its dependencies are
installed in a library that is readable on startup (e.g., the user
or site lib).  If it is not installed, then this script will fail.
}

