% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_packages}
\alias{install_packages}
\alias{install_packages_missing}
\title{Install packages}
\usage{
install_packages(packages, sources = package_sources(), lib = NULL, ...,
  error = TRUE, move_in_place = FALSE)

install_packages_missing(packages, ...)
}
\arguments{
\item{packages}{A character vector of packages to install}

\item{sources}{Source information from
\code{\link{package_sources}}.}

\item{lib}{Optional library to install packages into}

\item{...}{Additional arguments passed through to
\code{\link{install.packages}} (or for
\code{install_packages_missing}, to \code{install_packages}).}

\item{error}{Throw an error if package installation fails.}

\item{move_in_place}{Attempt to do the right thing during parallel
installations that affect the same packages.  When \code{TRUE},
installation happens into a temporary directory and then files
are copied into place.  This means that if multiple processes
are installing the same packages they should not get into a
tangle; they'll both install and compile everything (which is
much simpler than trying to negotiate lockfiles and waiting
periods) but the processes will not get into trouble where one
is disallowed from removing the packages used by another.  If
installation fails catastrophically this may leave some files in
\code{.libPaths()[[1]]} beginning \code{CONTEXT_TMP_LIBRARY_};
these can be safely removed.}
}
\description{
Install packages, using source information from
\code{\link{package_sources}}.  In contrast with
\code{install.packages} this function can install packages in a
new local repository (used to isolate packages from the rest of
the system) and can install packages originally found on github
and bitbucket by creating a transient drat repository.  In
addition, \code{install_packages_missing} only installs
\emph{missing} packages only.  Also in contrast with
\code{install.packages}, these function will throw an error if
package installation fails (unless error=FALSE is set).
}

