% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{is_absolute_path}
\alias{is_absolute_path}
\alias{is_relative_path}
\title{Test if path is absolute or relative}
\usage{
is_absolute_path(path)

is_relative_path(path)
}
\arguments{
\item{path}{Vector of paths to test}
}
\description{
Test if path is absolute or relative.
}

