% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{khodakA}
\alias{khodakA}
\title{Khodak et al's transformation matrix for GloVe}
\format{
A 300 by 300 matrix.
}
\source{
\url{https://github.com/NLPrinceton/ALaCarte}
}
\usage{
khodakA
}
\description{
A square matrix corresponding to the transformation matrix computed by Khodak et al.
}
\keyword{datasets}
