% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_propagate.R
\name{label_propagate}
\alias{label_propagate}
\title{Generate and select additional semantically related words from topical seed word(s)}
\usage{
label_propagate(seed_words, pre_trained, graphics = FALSE, N = 10, norm = "l2")
}
\arguments{
\item{seed_words}{(character) vector - seed word(s)}

\item{pre_trained}{(numeric) - a F x D matrix corresponding to pretrained embeddings.
F = number of features and D = embedding dimensions.}

\item{graphics}{(logical) - opens window for selection of words}

\item{N}{(numeric) - number of words to generate for selection}

\item{norm}{character = c("l2", "none") - how to scale input matrices. If they are already scaled - use "none" (see ?sim2)}
}
\value{

}
\description{
Generate and select additional semantically related words from topical seed word(s)
}
\examples{
nns_selected <- label_propagate(seed_words =  c("president", "executive"),
                                pre_trained = cr_glove_subset, N = 10,
                                graphics = T)
}
