% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{last_loaded_context}
\alias{last_loaded_context}
\title{Return last loaded context}
\usage{
last_loaded_context(error = TRUE)
}
\arguments{
\item{error}{Throw an error if no context has been loaded}
}
\description{
Return the last loaded context
}

