% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-ct.r
\name{latex_ct}
\alias{latex_ct}
\title{Print a cross-table to be copy-pasted or sourced into a TeX file.}
\usage{
latex_ct(x)
}
\arguments{
\item{x}{Numeric binary matrix.}
}
\value{
Nothing.
}
\description{
See \url{http://tex.stackexchange.com/a/32690} and
\url{http://tex.stackexchange.com/questions/49746/a-table-with-square-cells}.
}
\examples{
mat <- matrix(rbinom(size = 1, n = 7 * 12, p = .333), nrow = 7, ncol = 12)
latex_ct(x = mat)
colnames(mat) <- paste0("A", 1:ncol(mat))
latex_ct(x = mat)
}

