% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-scale-grep.r
\name{make_scale_grep}
\alias{make_scale_grep}
\title{Build conceptual scale based on pattern matching}
\usage{
make_scale_grep(objs, patterns, split)
}
\arguments{
\item{objs}{Vector (usually character or factor).}

\item{patterns}{Character vector of regular expressions. Each attribute of 
the resulting context amounts to \code{grepl(x = objs, pattern = pattern)} 
for some \code{pattern} in \code{patterns}.}

\item{split}{Character. Regular expression used to split \code{objs} into
\code{patterns}, if \code{patterns} is not provided.}
}
\value{
A binary matrix (formal context).
}
\description{
Build conceptual scale based on pattern matching
}

