% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-scale-regex.r
\name{make_scale_regex}
\alias{make_scale_regex}
\title{Build conceptual scale based on pattern matching}
\usage{
make_scale_regex(objs, patterns, split)
}
\arguments{
\item{objs}{Vector (usually character or factor).}

\item{patterns}{Character vector of regular expressions. Each attribute of 
the resulting context amounts to \code{grepl(x = objs, pattern = pattern)} 
for some \code{pattern} in \code{patterns}.}

\item{split}{Character. Regular expression used to split \code{objs} into
\code{patterns}, if \code{patterns} is not provided.}
}
\value{
A binary matrix (formal context).
}
\description{
Build conceptual scale based on pattern matching
}
\examples{
# state name modifiers
make_scale_regex(objs = state.name,
                 patterns = c("^New", "^(North|South|East|West)"))

# letters in state names
make_scale_regex(objs = state.abb, split = "")
}

