% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-chisq.r
\name{matrix_chisq}
\alias{matrix_chisq}
\title{Convert a frequency table (non-negative numeric matrix) to a matrix of the
same dimensions containing at each position \code{(i, j)} the chi-square
value associated with the 2-by-2 frequency table obtained by collapsing the
original matrix along all rows and columns other than \code{i} and \code{j}.}
\usage{
matrix_chisq(x)
}
\arguments{
\item{x}{Numeric non-negative matrix.}
}
\description{
Convert a frequency table (non-negative numeric matrix) to a matrix of the
same dimensions containing at each position \code{(i, j)} the chi-square
value associated with the 2-by-2 frequency table obtained by collapsing the
original matrix along all rows and columns other than \code{i} and \code{j}.
}

