% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-ratio.r
\name{matrix_ratio}
\alias{matrix_ratio}
\title{Convert a frequency table (non-negative numeric matrix) to a matrix of the 
same dimensions containing at each position \code{(i, j)} the ratio of the
joint probability $P_{(i,j)}$ to the product $P_iP_j$ of the marginal
probabilities (based on the row and column sums).}
\usage{
matrix_ratio(x)
}
\arguments{
\item{x}{Numeric non-negative matrix.}
}
\description{
Convert a frequency table (non-negative numeric matrix) to a matrix of the 
same dimensions containing at each position \code{(i, j)} the ratio of the
joint probability $P_{(i,j)}$ to the product $P_iP_j$ of the marginal
probabilities (based on the row and column sums).
}

