% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nns.R
\name{nns}
\alias{nns}
\title{Given a set of embeddings and a set of candidate neighbors, find the top N nearest
neighbors.}
\usage{
nns(x, N = 10, candidates = character(0), pre_trained)
}
\arguments{
\item{x}{a \linkS4class{dem} or \linkS4class{fem} object}

\item{N}{(numeric) number of nearest neighbors to return}

\item{candidates}{(character) vector of features to consider as candidates to be nearest neighbor}

\item{pre_trained}{a F x D matrix of numeric values corresponding to pretrained embeddings
F = number of features and D = embedding dimensions.
rownames(pre_trained) = set of features for which there is a pre-trained embedding}
}
\value{
a \code{data.frame} with following columns:
\item{\code{target}}{ (character) vector with the rownames of the dfm,
either defining the groups or the target terms}.
\item{\code{feature}}{(character) vector of features from the candidate set,
one instance for each target.}
\item{\code{value}}{(numeric) cosine similarity between target
and candidate.}
}
\description{
Given a set of embeddings and a set of candidate neighbors, find the top N nearest
neighbors.
}
\examples{

library(quanteda)

# tokenize text
anes2016_toks <- tokens(anes2016_sample_corpus)

# construct document-feature-matrix
anes2016_dfm <- dfm(anes2016_toks)

# construct document-embedding-matrix
anes2016_dem <- dem(anes2016_dfm, pre_trained = glove_subset,
transform = TRUE, transform_matrix = khodakA, verbose = TRUE)

# group document-embedding-matrix
anes2016_dem_ideology <- dem_group(anes2016_dem,
groups = anes2016_dem@docvars$ideology)

# find nearest neighbors
nns(x = anes2016_dem_ideology,
pre_trained = glove_subset, candidates = NULL, N = 10)
}
\keyword{nns}
