% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nns_ratio.R
\name{nns_ratio}
\alias{nns_ratio}
\title{Computes the ratio of cosine similarities for two embeddings over
the union of their respective top N nearest neighbors.}
\usage{
nns_ratio(
  x,
  N = 10,
  numerator = NULL,
  candidates = character(0),
  pre_trained,
  verbose = TRUE
)
}
\arguments{
\item{x}{a \linkS4class{dem} or \linkS4class{fem} object or in general a matrix of embeddings}

\item{N}{number of nearest neighbors to consider}

\item{numerator}{(character) defines which group is the nuemerator in the ratio
Note: if nearest neighbors overlap, the resulting number of features
will be fewer than 2*N.}

\item{candidates}{character vector delimiting the features to consider as nearest neighbor candidates}

\item{pre_trained}{a F x D matrix of numeric values corresponding to pretrained embeddings
F = number of features and D = embedding dimensions.
rownames(pre_trained) = set of features for which there is a pre-trained embedding}

\item{verbose}{provide information on which group is the numerator}
}
\value{
a \code{data.frame} with following columns:
\item{\code{feature}}{(character) vector of features from the candidate set,
one instance for each target.}
\item{\code{value}}{(numeric) ratio of cosine similarities}.
}
\description{
Computes the ratio of cosine similarities for two embeddings over
the union of their respective top N nearest neighbors.
}
\examples{

library(quanteda)

# tokenize corpus
toks <- tokens(cr_sample_corpus)

# build a tokenized corpus of contexts sorrounding a target term
immig_toks <- tokens_context(x = toks, pattern = "immigr*", window = 6L)

# build document-feature matrix
immig_dfm <- dfm(immig_toks)

# construct document-embedding-matrix
immig_dem <- dem(immig_dfm, pre_trained = cr_glove_subset,
transform = TRUE, transform_matrix = cr_transform, verbose = FALSE)

# to get group-specific embeddings, average within party
immig_wv_party <- dem_group(immig_dem, groups = immig_dem@docvars$party)

# compute the cosine similarity between each party's
# embedding and a specific set of features
nns_ratio(x = immig_wv_party, N = 10, numerator = "R",
candidates = immig_wv_party@features,
pre_trained = cr_glove_subset, verbose = FALSE)
}
\keyword{nns_ratio}
