% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{open}
\alias{open}
\title{Open file implementation using a context manager}
\usage{
open(file_name, mode = "r")
}
\arguments{
\item{file_name}{File name to open.}

\item{mode}{File mode.}
}
\value{
S3 class ContextManager
}
\description{
Similar to Python's `open()`, but creates a ContextManager S3 object.
Intended to be used as a first (left) argument of `%as%`, or as any but last
argument of `context::with()`.
}
