% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{open_file}
\alias{open_file}
\title{Open file implementation using a context manager}
\usage{
open_file(file_name, mode, as)
}
\arguments{
\item{file_name}{File name to open.}

\item{mode}{File mode.}

\item{as}{Character specifying the name of the variable.}
}
\value{
S3 class ContextManager
}
\description{
Similar to Python's `open()`, intended to be used with a context manager.
}
