% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drat.R
\name{package_sources}
\alias{package_sources}
\title{Information about non-CRAN packages}
\usage{
package_sources(cran = NULL, repos = NULL, github = NULL,
  bitbucket = NULL, local = NULL, expire = 1)
}
\arguments{
\item{cran}{Use this to set the CRAN mirror.  The default will use
the Rstudio CRAN mirror which is usually nice and fast.}

\item{repos}{Character vector of additional reposotories to use.
Specify whatever URL would normally work with
\code{\link{install.packages}}.  In addition, specify URL of the
form \code{drat://<username>} to install from drat repositories.}

\item{github, bitbucket}{Character vector of github or bitbucket
packages to install.  The format is similar to
\code{devtools::install_github}, being
\code{username/repo[/subdir][@ref]} (note that pull requests are
not supported in contrast with devtools).}

\item{local}{Character vector of paths to packages on your local
computer to make available.  These can be directories or built
\code{tar.gz}, \code{tgz} or \code{zip} files.  If the directory
contains a \code{DESCRIPTION} file it is treated as the source
of an R package, otherwise all package-like files (tar.gz, tgz
and zip) will be added.}

\item{expire}{Number of days after which to expire packages in the
archive.  After this, packages will be fetched again.
Fractional days or \code{\link{difftime}} objects are fine.}
}
\description{
Specify information about locations to find non-CRAN packages.
This will typically be saved with a \code{context}.  As
part of the context generation process, any packages not in a
repository (CRAN or other CRAN-like repository such as a drat
reposotory) will be copied into a local drat repository that will
be available to a process that is restoring the context.
}

