% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parallel_cluster_start}
\alias{parallel_cluster}
\alias{parallel_cluster_start}
\alias{parallel_cluster_stop}
\title{Start a sub-cluster}
\usage{
parallel_cluster_start(n, ctx)

parallel_cluster_stop()

parallel_cluster()
}
\arguments{
\item{n}{The number of nodes.  No attempt at guessing this number
is made as that is terribly error prone.  If you're using this
function you should know how many resources you have available.}

\item{ctx}{The context to initialise on each cluster node.}
}
\description{
Start a sub-cluster, using the \code{parallel} package.  This will
be available via either the return value of this function, the
\code{parallel_cluster} function or by using \code{cl = NULL} with
any of the \code{parallel} package functions.  The cluster will be
started so that it is ready to use the context.
}

