% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conText.R
\name{permute_ols}
\alias{permute_ols}
\title{Permute OLS}
\usage{
permute_ols(Y = NULL, X = NULL)
}
\arguments{
\item{Y}{vector of regression model's dependent variable (embedded context)}

\item{X}{data.frame of model independent variables (covariates)}
}
\value{
list with two elements, \code{betas} = list of beta_coefficients (D dimensional vectors);
\code{normed_betas} = tibble with the norm of the non-intercept coefficients
}
\description{
Estimate empirical p-value using permutated regression
}
