% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nns_ratio.R
\name{plot_nns_ratio}
\alias{plot_nns_ratio}
\title{Plot output of get_nns_ratio}
\usage{
plot_nns_ratio(x, alpha = 0.01, horizontal = TRUE)
}
\arguments{
\item{x}{output of get_nns_ratio}

\item{alpha}{(numerical) betwee 0 and 1;significance threshold to identify significant values}

\item{horizontal}{(logical) defines the type of plot}
}
\value{
a ggplot
}
\description{
Plot output of get_nns_ratio
}
\examples{

library(ggplot2)
library(quanteda)

# tokenize corpus
toks <- tokens(cr_sample_corpus)

# build a tokenized corpus of contexts sorrounding a target term
immig_toks <- tokens_context(x = toks, pattern = "immigr*", window = 6L)

# we limit candidates to features in our corpus
feats <- featnames(dfm(immig_toks))

# compute ratio
set.seed(2021L)
immig_nns_ratio <- get_nns_ratio(x = immig_toks,
                                 N = 10,
                                 groups = docvars(immig_toks, 'party'),
                                 numerator = "R",
                                 candidates = feats,
                                 pre_trained = cr_glove_subset,
                                 transform = TRUE,
                                 transform_matrix = cr_transform,
                                 bootstrap = TRUE,
                                 num_bootstraps = 10,
                                 permute = TRUE,
                                 num_permutations = 10,
                                 verbose = FALSE)

plot_nns_ratio(x = immig_nns_ratio, alpha = 0.01, horizontal = TRUE)
}
\keyword{plot_nns_ratio}
