% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provision.R
\name{provision_context}
\alias{provision_context}
\title{Provision a context}
\usage{
provision_context(ctx, platform = NULL, version = NULL, quiet = FALSE,
  allow_missing = FALSE, installed_action = "skip",
  additional_libraries = NULL, refresh_drat = FALSE)
}
\arguments{
\item{ctx}{A context object.  We'll look in here for packages to install.}

\item{platform}{Platform to create the library for}

\item{version}{Version of R to create the library for}

\item{quiet}{Be quiet when installing source packages}

\item{allow_missing}{Allow packages requiring compilation to be
skipped over?  If \code{TRUE} you'll need to sort these out
yourself.}

\item{installed_action}{Action if packages are installed}

\item{additional_libraries}{Character vector of additional
libraries to pass through to \code{provisionr}.  Packages here
will be counted when determining what to install and upgrade
(based on \code{installed_action}).}

\item{refresh_drat}{Force-refresh any local packages that are
listed in the \code{package_sources} argument to
\code{context_save}}
}
\description{
Provision a context
}

