% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_line}
\alias{read_line}
\title{Read a single line from file connectoin}
\usage{
read_line(connection)
}
\arguments{
\item{connection}{File connection.}
}
\value{
Character string of one line in a file.
}
\description{
Read a single line from file connectoin
}
