% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_test_fastq}
\alias{read_test_fastq}
\title{Path to a small simulated fastq file}
\usage{
read_test_fastq()
}
\value{
A character vector with the file path.
}
\description{
Path to a small simulated fastq file
}
