% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{restore_locals}
\alias{restore_locals}
\title{Restore locals}
\usage{
restore_locals(dat, parent, db)
}
\arguments{
\item{dat}{An expression that has been through
\code{prepare_expression}.  Key elements are
\code{function_hash} and \code{objects}}

\item{parent}{The parent environment to restore locals to}

\item{db}{The database used to prepare the expression}
}
\description{
Restore locals created by \code{\link{prepare_expression}}.
}
