% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conText.R
\name{run_jack_ols}
\alias{run_jack_ols}
\title{Run jackknife debiased OLS}
\usage{
run_jack_ols(X, Y, confidence_level = 0.95)
}
\arguments{
\item{X}{data.frame of model independent variables (covariates)}

\item{Y}{vector of regression model's dependent variable (embedded context)}

\item{confidence_level}{(numeric in (0,1)) confidence level e.g. 0.95}
}
\value{
list with two elements, \code{betas} = list of beta_coefficients (D dimensional vectors);
\code{normed_betas} = tibble with the norm and CIs of the non-intercept coefficients
}
\description{
Run jackknife debiased OLS
}
