% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conText.R
\name{run_ols}
\alias{run_ols}
\title{Run OLS}
\usage{
run_ols(Y = NULL, X = NULL, ids = NULL)
}
\arguments{
\item{Y}{vector of regression model's dependent variable (embedded context)}

\item{X}{data.frame of model independent variables (covariates)}

\item{ids}{vector of cluster ids}
}
\value{
list with two elements, \code{betas} = list of beta_coefficients (D dimensional vectors);
\code{normed_betas} = tibble with the norm of the non-intercept coefficients
}
\description{
Bootstrap model coefficients and standard errors
}
