% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{run_task}
\alias{run_task}
\title{Run a task}
\usage{
run_task(handle, install = FALSE, envir = .GlobalEnv)
}
\arguments{
\item{handle}{Task handle}

\item{install}{Install packages when constructing context?}

\item{envir}{Environment to load global variables into.}
}
\description{
Run a task
}

