% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_corpus}
\alias{sample_corpus}
\title{Sample corpus}
\format{
A data frame with 2000 rows and 2 variables:
\describe{
\item{speech}{words spoken by a given representative}
\item{party}{party membership of said representative}
...
}
}
\source{
\url{https://data.stanford.edu/congress_text}
}
\usage{
sample_corpus
}
\description{
A sample from the congressional record corpus
1000 speeches from each of the two major parties, sessions 111th - 114th
}
\keyword{datasets}
