% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_glove}
\alias{sample_glove}
\title{Sample GloVe}
\format{
A double object with 70 rows and 300 columns:
\describe{
\item{row}{each row corresponds to a word}
\item{column}{each column corresponds to a dimension in the embedding space}
...
}
}
\source{
\url{https://nlp.stanford.edu/projects/glove/}
}
\usage{
sample_glove
}
\description{
A subset of Stanford GloVe embeddings.
}
\keyword{datasets}
