% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-by-predicate.r
\name{scale_by_predicate}
\alias{scale_by_predicate}
\title{Expand variables in a data frame via logical conceptual scaling}
\usage{
scale_by_predicate(data, predicates)
}
\arguments{
\item{data}{Data frame.}

\item{predicates}{A logical expression involving variables in \code{data}, or
a list of such expressions. If a list, then the names of any named elements
will be used as column names in the resulting context.}
}
\value{
A binary numeric matrix (formal context) of \code{nrow(data)} rows 
  and 1 column, whose entries are the truth values of \code{predicate} 
  applied to the rows of \code{data}.
}
\description{
Expand variables in a data frame via logical conceptual scaling
}
\examples{
# logical expressions involving single variables
scale_by_predicate(mtcars, cyl == 8)
scale_by_predicate(mtcars, mpg > 20)

# logical expressions involving multiple variables
scale_by_predicate(mtcars, cyl > 4 & wt < 3)

# multiple logical expressions
scale_by_predicate(mtcars, list(cyl > 4, wt < 3, mpg > 20))

# singleton list
scale_by_predicate(mtcars, list(cyl == 6))

# list with some predicates named
scale_by_predicate(mtcars, list(`4-cyl` = cyl == 4,
                                wt > 2,
                                goodMPG = mpg > 20))
}

