% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-by-variable.r
\name{scale_by_variable}
\alias{scale_by_variable}
\alias{scale_by_variables}
\title{Expand variables in a data frame via contextualized or named conceptual
scaling}
\usage{
scale_by_variable(data, var, scale)

scale_by_variables(data, vars = names(scales), scales)
}
\arguments{
\item{data}{Data frame.}

\item{var}{Character. The name of the column of \code{data} that \code{scale}
scales.}

\item{scale}{Binary numeric matrix or character. The scale to use to 
transform \code{var}, either as a formal context (binary matrix) or as a 
standard type (passed to \code{\link{make_scale_standard}}).}

\item{vars}{Character vector. The names of the columns of \code{data} that
\code{scales} scale.}

\item{scales}{List having the same length as \code{vars} containing the
scales to transform \code{vars}.}
}
\value{
A binary matrix (formal context).
}
\description{
Expand variables in a data frame via contextualized or named conceptual
scaling
}
\examples{
titanic_data <- as.data.frame(Titanic)

# custom scale for Class variable
class_scale <- rbind(
  c(1, 1, 1, 0),
  c(0, 1, 1, 0),
  c(0, 0, 1, 0),
  c(0, 1, 1, 1)
)
rownames(class_scale) <- levels(titanic_data$Class)
colnames(class_scale) <- paste0(">", levels(titanic_data$Class))
print(class_scale)
scale_by_variable(data = titanic_data, var = "Class", scale = class_scale)

# nominal scale for Sex variable
scale_by_variable(data = titanic_data, var = "Sex", scale = "nominal")

# ordinal scale for Age variable
scale_by_variable(data = titanic_data, var = "Age", scale = "ordinal")

# multiple scales applied at once
both_vars <- c("Class", "Sex")
sex_scale <- make_scale_standard(objs = titanic_data$Sex, scale = "nominal")
both_scales <- list(class_scale, sex_scale)
scale_by_variables(data = titanic_data, vars = both_vars, scales = both_scales)
}

