% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_context.r
\name{scale_context}
\alias{scale_context}
\alias{scales_context}
\title{Expand variables in a data frame via conceptual scaling}
\usage{
scale_context(data, var, scale)

scales_context(data, vars = names(scales), scales)
}
\arguments{
\item{data}{Data frame.}

\item{var}{Character. The name of the column of \code{data} that \code{scale}
scales.}

\item{scale}{Binary numeric matrix or character. The scale to use to
transform \code{var}, either as a formal context (binary matrix) or as a
standard type (passed to \code{\link{scale_standard}}).}

\item{vars}{Character vector. The names of the columns of \code{data} that
\code{scales} scale.}

\item{scales}{List having the same length as \code{vars} containing the
scales to transform \code{vars}.}
}
\value{
A binary matrix (formal context).
}
\description{
Expand variables in a data frame via conceptual scaling
}
\examples{
titanic_data <- as.data.frame(Titanic)

class_scale <- rbind(
  c(1, 1, 1, 0),
  c(0, 1, 1, 0),
  c(0, 0, 1, 0),
  c(0, 1, 1, 1)
)
rownames(class_scale) <- levels(titanic_data$Class)
colnames(class_scale) <- paste0(">", levels(titanic_data$Class))

scale_context(data = titanic_data, var = "Class", scale = class_scale)

scale_context(data = titanic_data, var = "Sex", scale = "nominal")

scale_context(data = titanic_data, var = "Age", scale = "ordinal")

both_vars <- c("Class", "Sex")

sex_scale <- rbind(
  c(1, 0),
  c(0, 1)
)
rownames(sex_scale) <- levels(titanic_data$Sex)
colnames(sex_scale) <- c("Male", "Female")

both_scales <- list(class_scale, sex_scale)

scales_context(data = titanic_data, vars = both_vars, scales = both_scales)
}

