% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_standard.r
\name{scale_standard}
\alias{scale_biordinal}
\alias{scale_boolean}
\alias{scale_interordinal}
\alias{scale_nominal}
\alias{scale_ordinal}
\alias{scale_revordinal}
\alias{scale_standard}
\title{Standard scales}
\usage{
scale_standard(objs, x, n, scale = "nominal", ...)

scale_nominal(objs)

scale_boolean(objs)

scale_ordinal(objs)

scale_revordinal(objs)

scale_interordinal(objs)

scale_biordinal(objs, cut = ceiling(length(objs)/2))
}
\arguments{
\item{objs}{Vector (usually character or factor).}

\item{x}{Factor vector. Ignored if \code{objs} is present.}

\item{n}{Positive integer. The number of objects in the scale. Ignored if 
\code{objs} or \code{var} is present.}

\item{scale}{Character. Name of standard scale, e.g. "ordinal". Defaults to
"nominal".}

\item{...}{Additional arguments (listed below) passed from 
\code{scale_standard} to specific scale generators.}

\item{cut}{Numeric. Integer between 0 and \code{length(objs)} indicating the 
index at which to cut \code{objs} for biordinal scaling.}
}
\description{
Standard scales
}
\examples{
titanic_data <- as.data.frame(Titanic)

# nominal scale
print(scale_nominal(objs = unique(titanic_data$Class)))

# ordinal scale
print(scale_ordinal(objs = unique(titanic_data$Class)))

# other scales
for (scale in c("boolean", "revordinal", "interordinal")) {
  print(scale_standard(x = titanic_data$Class, scale = scale))
}

# biordinal scale with custom cut value
print(scale_biordinal(objs = unique(titanic_data$Class), cut = 3))
}

