% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_query.R
\name{task_context_id}
\alias{task_context_id}
\title{Find context for a task}
\usage{
task_context_id(ids, db)
}
\arguments{
\item{ids}{Vector of task ids}

\item{db}{Something that can be converted to a context db object
(a database, root or context).}
}
\description{
Find the context id associated with a task
}
