% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_query.R
\name{task_deps}
\alias{task_deps}
\title{Task dependencies}
\usage{
task_deps(ids, db, named = FALSE)
}
\arguments{
\item{ids}{Vector of task ids}

\item{db}{Something that can be converted to a context db object
(a database, root or context).}

\item{named}{Name the output with the task ids?}
}
\description{
Task dependencies
}
