% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_query.R
\name{task_exists}
\alias{task_exists}
\alias{task_list}
\title{List tasks}
\usage{
task_exists(ids, db)

task_list(db)
}
\arguments{
\item{ids}{Vector of task ids}

\item{db}{Something that can be converted to a context db object
(a database, root or context).}
}
\description{
List tasks and test if they exist
}
