% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_expr}
\alias{task_expr}
\alias{task_function_name}
\title{Fetch task expression}
\usage{
task_expr(handle, locals = FALSE)

task_function_name(handle)
}
\arguments{
\item{handle}{A task handle}

\item{locals}{Return locals bound to the expression (as an
attribute "locals")}
}
\description{
Fetch expression for a task
}

