% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_query.R
\name{task_function_name}
\alias{task_function_name}
\title{Fetch task function name}
\usage{
task_function_name(ids, db)
}
\arguments{
\item{ids}{Vector of task ids}

\item{db}{Something that can be converted to a context db object
(a database, root or context).}
}
\description{
Fetch function name for a task
}

