% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_log}
\alias{task_log}
\title{Return task log}
\usage{
task_log(root, id)
}
\arguments{
\item{root}{Context root (or a context, or anything context can
internally wrangle into its database format; see
\code{\link{context_db}}).}

\item{id}{Task identifier}
}
\description{
Return the log of a task, if enabled
}

