% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_run.R
\name{task_reset}
\alias{task_reset}
\title{Reset status and submission time of tasks}
\usage{
task_reset(id, context)
}
\arguments{
\item{id}{A vector of task identifiers}

\item{context}{A context object}
}
\description{
Reset tasks
}
