% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_query.R
\name{task_result}
\alias{task_result}
\title{Fetch task result}
\usage{
task_result(id, db, allow_incomplete = FALSE)
}
\arguments{
\item{id}{Single task identifier}

\item{db}{Something that can be converted to a context db object
(a database, root or context).}

\item{allow_incomplete}{Should we avoid throwing an error if a task is not
completed?  Used internally, and not generally needed.}
}
\description{
Fetch result from completed task.
}

