% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_run_external}
\alias{task_run_external}
\title{Run a task in separate process}
\usage{
task_run_external(root, identifier, task_id, path_log)
}
\arguments{
\item{root}{Something interpetable as the context root; either}

\item{identifier}{Either the id or name of a context (see
\code{\link{context_list}})}

\item{task_id}{A task identifier}

\item{path_log}{Path to log file}
}
\description{
Run a task in a separate process, using [callr::r]. Unlike
[context::task_run] this does not return the value, and is called
for the side effect of writing to the context.
}
\author{
Richard Fitzjohn
}
