% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_query.R
\name{task_times}
\alias{task_times}
\title{Fetch task times}
\usage{
task_times(ids, db, unit_elapsed = "secs", sorted = TRUE)
}
\arguments{
\item{ids}{Vector of task ids}

\item{db}{Something that can be converted to a context db object
(a database, root or context).}

\item{unit_elapsed}{Elapsed time unit.  The default is "secs".
This is passed to the \code{as.numeric} method of a
\code{difftime} object.}

\item{sorted}{Sort the output in terms of submitted time?  If
\code{FALSE} then the output is sorted based on task ids.}
}
\description{
Fetch times taken to queue, run, and since running a task.
}
\author{
Rich FitzJohn
}
