% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{tasks_delete}
\alias{tasks_delete}
\title{Delete a task}
\usage{
tasks_delete(id, root)
}
\arguments{
\item{handle}{A task handle}
}
\value{
\code{TRUE} if a task was actually deleted.
}
\description{
Delete a task, including its results.
}

