% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_query.R
\name{tasks_status}
\alias{tasks_status}
\title{Task status}
\usage{
tasks_status(id, root, named = FALSE)
}
\arguments{
\item{named}{Name the output with the task ids?}

\item{handle}{Task handle}
}
\description{
Task status
}

