% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{with.ContextAs}
\alias{with.ContextAs}
\title{Context manager's `with` function}
\usage{
\method{with}{ContextAs}(data, expr, ...)
}
\arguments{
\item{data}{[`ContextAs`] S3 object with elements `context` and `variable`}

\item{expr}{Expression or code block to be evaluated within the context}

\item{...}{Unused. Here for consistency with `base::with()`.}
}
\description{
Context manager's `with` function
}
