% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{with.list}
\alias{with.list}
\title{Context manager's `with` function}
\usage{
\method{with}{list}(data, expr, ...)
}
\arguments{
\item{data}{List of `ContextAs` S3 objects.}

\item{expr}{Expression or code block to be evaluated within the context}

\item{...}{Unused. Here for consistency with `base::with()`.}
}
\description{
Context manager's `with` function
}
