% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{with}
\alias{with}
\title{Context manager's `with` function}
\usage{
with(context, expression)
}
\arguments{
\item{context}{[`ContextManager`] object}

\item{expression}{Expression or code block to be evaluated within the context}
}
\description{
Context manager's `with` function
}
