% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{with_}
\alias{with_}
\title{Context manager's `with` function}
\usage{
with_(...)
}
\arguments{
\item{...}{ContextAs S3 objects, followed by an expression.}
}
\description{
Context manager's `with` function
}
