% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-context.r
\name{write_context}
\alias{write_context}
\alias{write_context_concepts}
\alias{write_context_coron}
\title{Write binary matrices to cross-table files.}
\usage{
write_context(x, file = "", format)

write_context_coron(x, file = "")

write_context_concepts(x, file = "")
}
\arguments{
\item{x}{Numeric binary matrix, preferably with row and column names.}

\item{file}{Character. File name (with or without extension) to export formal
context to.}

\item{format}{Character. Formatting to use, from among "coron"/"rcf" and 
"concepts".}
}
\value{
Nothing.
}
\description{
Write binary matrices to cross-table files.
}

