% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-ct.r
\name{write_ct}
\alias{write_ct}
\title{Write binary matrices to cross-table files (.csv for the concepts module in
Python).}
\usage{
write_ct(x, file = "")
}
\arguments{
\item{x}{Numeric binary matrix, preferably with row and column names.}

\item{file}{Character. File name (with path, preferably ending in 
\code{.csv}) to export formal context to.}
}
\value{
Nothing.
}
\description{
Write binary matrices to cross-table files (.csv for the concepts module in
Python).
}

