% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-rcf.r
\name{write_rcf}
\alias{write_rcf}
\title{Write binary matrices to cross-table files (\code{.rcf} for Coron).}
\usage{
write_rcf(x, name = "Default Name", file = "")
}
\arguments{
\item{x}{Numeric binary matrix, preferably with row and column names.}

\item{name}{Character. Name given to the formal context, included in the
header to the \code{.rcf} file.}

\item{file}{Character. File name (with path, preferably ending in
\code{.rcf}) to export formal context to.}
}
\value{
Nothing.
}
\description{
Write binary matrices to cross-table files (\code{.rcf} for Coron).
}

