% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_nns.R
\name{bootstrap_nns}
\alias{bootstrap_nns}
\title{Bootstrap nearest neighbors}
\usage{
bootstrap_nns(
  context = NULL,
  pre_trained = NULL,
  transform_matrix = NULL,
  transform = TRUE,
  candidates = NULL,
  bootstrap = TRUE,
  num_bootstraps = 20,
  N = 50,
  norm = "l2"
)
}
\arguments{
\item{context}{vector of texts - `context`` variable in get_context output}

\item{pre_trained}{a V x D matrix of numeric values - pretrained embeddings with V = size of vocabulary and D = embedding dimensions}

\item{transform_matrix}{a D x D transformation matrix}

\item{transform}{logical - if TRUE (default) apply the a la carte transformation, if FALSE ouput untransformed averaged embedding}

\item{candidates}{character vector defining the candidates for nearest neighbors - e.g. output from \code{get_local_vocab}}

\item{bootstrap}{logical - if TRUE, bootstrap cosine similarity - required to get standard errors cosine similarity of nearest neighbors}

\item{num_bootstraps}{numeric - number of bootstraps to use}

\item{N}{number of nearest neighbors to return}

\item{norm}{character = c("l2", "none") - set to 'l2' for cosine similarity and to 'none' for inner product (see ?sim2 in text2vec)}
}
\value{
data.frame with nearest neighbor tokens, similarity (numeric value) and std.error (given bootstrap)
}
\description{
Bootstrap nearest neighbors
}
\examples{
library(conText)
library(dplyr)

# load data
corpus <- sample_corpus
pre_trained <- sample_glove
transform_matrix <- khodakA

# find contexts of immigration
context_immigration <- get_context(x = corpus$speech,
                                   target = 'immigration',
                                   window = 6,
                                   valuetype = "fixed",
                                   case_insensitive = TRUE,
                                   hard_cut = FALSE, verbose = FALSE)

# find local vocab (use it to define the candidate of nearest neighbors)
local_vocab <- get_local_vocab(context_immigration$context, pre_trained)

set.seed(42L)
nns_immigration <- bootstrap_nns(context = context_immigration$context,
                                 pre_trained, transform_matrix,
                                 transform = TRUE, candidates = local_vocab,
                                 bootstrap = TRUE,
                                 num_bootstraps = 20, N = 50, norm = "l2")

# print output
knitr::kable(head(nns_immigration))
}
