% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-builder.R
\name{build_fem}
\alias{build_fem}
\title{build a \code{fem-class} object}
\usage{
build_fem(
  Class = "fem",
  x_fem,
  features = character(),
  counts = numeric(),
  Dimnames = list()
)
}
\arguments{
\item{Class}{defines the class of this object (fixed)}

\item{x_fem}{a \verb{dgCMatrix class} matrix}

\item{features}{features used in computing the embeddings}

\item{counts}{counts of features used in computing embeddings}

\item{Dimnames}{row (features) and columns (NULL) names}
}
\description{
build a \code{fem-class} object
}
