% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_transform.R
\name{compute_transform}
\alias{compute_transform}
\title{compute transformation matrix A}
\usage{
compute_transform(x, pre_trained, weighting = 500)
}
\arguments{
\item{x}{a \code{fcm}}

\item{pre_trained}{a \code{dgTMatrix-class} matrix of word vectors}

\item{weighting}{1 = no weighting; log = weight by the log of the frequency count; a numeric = threshold based weighting (= 1 if token count meets threshold, 0 ow); use log for small corpora, numeric threshold for larger corpora}
}
\value{
a \code{dgTMatrix-class} D x D (D = word vector dimensions)
}
\description{
compute transformation matrix A
}
\examples{

# note, cr_corpus_sample is too small to produce sensical word vectors
features <- prune_features(cr_corpus_sample, frequency_min = 5) \%>\% .$feature # subset features
cr_fcm <- create_fcm(x = cr_corpus_sample, window = 6L, target = features, features_keep = features) # create feature co-occurrence matrix
cr_glove <- compute_glove(x = cr_fcm, rank = 300, x_max = 10, n_iter = 10) # estimate local glove embeddings
cr_transform <- compute_transform(x = cr_fcm, pre_trained = cr_glove, weighting = 'log') # estimate local transform
}
\keyword{compute_transform}
