% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_transform.R
\name{compute_transform}
\alias{compute_transform}
\title{compute transformation matrix A}
\usage{
compute_transform(x, pre_trained, weighting = 500)
}
\arguments{
\item{x}{a \code{fcm}}

\item{pre_trained}{a \code{dgTMatrix-class} matrix of word vectors}

\item{weighting}{1 = no weighting; log = weight by the log of the frequency count; a numeric = threshold based weighting (= 1 if token count meets threshold, 0 ow); use log for small corpora, numeric threshold for larger corpora}
}
\value{
a \code{dgTMatrix-class} D x D (D = word vector dimensions)
}
\description{
compute transformation matrix A
}
\examples{

library(quanteda)

# note, cr_sample_corpus is too small to produce sensical word vectors

# tokenize
cr_toks <- tokens(cr_sample_corpus)

# construct feature-co-occurrence matrix
cr_fcm <- fcm(cr_toks, context = "window", window = 6,
count = "weighted", weights = 1 / (1:6), tri = FALSE)

# you will generally want to estimate a new (corpus-specific)
# GloVe model, we will use cr_glove_subset instead
# see the Quick Start Guide to see a full example.

# estimate transform
cr_subset_transform <- compute_transform(x = cr_fcm,
pre_trained = cr_glove_subset, weighting = 'log')
}
\keyword{compute_transform}
