% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_transform.R
\name{compute_transform}
\alias{compute_transform}
\title{compute transform matrix A}
\usage{
compute_transform(context_fcm, pre_trained, vocab, weighting = 500)
}
\arguments{
\item{context_fcm}{a quanteda symmetrical fcm}

\item{pre_trained}{matrix of numeric values - pretrained embeddings}

\item{vocab}{the output of text2vec's create_vocabulary (or an equivalent)}

\item{weighting}{none = all words are weighted by their frequency count; log = weight by the log of the frequency count; a numeric = threshold based weighting (= 1 if token count meets threshold, 0 ow); use log for small corpora, numeric threshold for larger corpora}
}
\value{
matrix of numeric values - the D by D transformation matrix (D = number of dimensions of the embeddings space)
}
\description{
compute transform matrix A
}
