% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context_save}
\alias{context_load}
\alias{context_read}
\alias{context_save}
\title{Save and reload contexts}
\usage{
context_save(packages = NULL, sources = NULL, auto = FALSE,
  package_sources = NULL, envir = parent.frame(), root = tempdir())

context_load(handle, install = TRUE, envir = .GlobalEnv, ...)

context_read(handle)
}
\arguments{
\item{packages}{A character vector of packages (or \code{NULL}) if
no packages are to be loaded.}

\item{sources}{A character vector of source files to load.}

\item{auto}{Attempt to create the context automatically.  In this
case, do not specify either \code{packages} or \code{sources}.}

\item{package_sources}{Optional information about where to find
non-CRAN packages.  See \code{\link{package_sources}}.}

\item{envir}{The current environment.  This is used to copy
\emph{local} enviroments around, as these are needed even in the
case of non-automatic contexts.  For \code{context_load} this
is the environment into which the global environment is copied.
Specify a non-global environment here to avoid clobbering the
workspace, but at the risk that some environments may not
restore exactly as desired.}

\item{root}{Root directory to store and retrieve files.  Files
will be added to the \code{contexts} subdirectory of this path.
This will change later to support alternative ways of saving
files.  The default puts files into the temp directory of
\emph{this} R instance; this directory will be deleted on
session exit so it not an appropriate place to store files for
later use.}

\item{handle}{A \code{context_handle} object returned by
\code{context_save}.}

\item{install}{Install missing packages?}

\item{...}{Additional arguments passed through to
\code{install_packages} if it is used.}
}
\description{
Save and reload contexts.  Contexts consist of packages and
sources, or (if \code{auto}) is \code{TRUE}) a copy of the current
global environment.  Environments saved with \code{context_save}
can be reloaded with \code{context_load}; note that doing this
will probably alter the search path by loading any number of
packages.
}
\details{
The \code{context_read} function simply reads the context, but
does not load it.  It is mostly useful for debugging.
}

