% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conText.R
\name{conText}
\alias{conText}
\title{Embedding regression}
\usage{
conText(
  formula,
  corpus,
  pre_trained,
  transform = TRUE,
  transform_matrix,
  features = character(0),
  bootstrap = TRUE,
  num_bootstraps = 20,
  stratify = TRUE,
  permute = TRUE,
  num_permutations = 100,
  window = 6,
  valuetype = "fixed",
  case_insensitive = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted}

\item{corpus}{a quanteda \code{corpus}}

\item{pre_trained}{a F x D matrix of numeric values corresponding to pretrained embeddings
F = number of features and D = embedding dimensions.
rownames(pre_trained) = set of features for which there is a pre-trained embedding}

\item{transform}{(logical) if TRUE (default) apply the \link{a la carte} transformation, if FALSE ouput untransformed averaged embedding}

\item{transform_matrix}{a D x D matrix of numeric values corresponding to the `a la carte`` transformation matrix}

\item{features}{(character) a character vector of features to use when computing embedding}

\item{bootstrap}{(logical) if TRUE, bootstrap regression - required to get standard errors for normed coefficients}

\item{num_bootstraps}{(numeric) number of bootstraps to use}

\item{stratify}{(logical) if TRUE, stratify by covariates when bootstrapping}

\item{permute}{(logical) if TRUE, compute empirical p-values using permutation test}

\item{num_permutations}{(numeric) number of permutations to use}

\item{window}{the number of context words to be displayed around the keyword}

\item{valuetype}{the type of pattern matching: \code{"glob"} for "glob"-style
wildcard expressions; \code{"regex"} for regular expressions; or \code{"fixed"} for
exact matching. See \link[quanteda]{valuetype} for details.}

\item{case_insensitive}{logical; if \code{TRUE}, ignore case when matching a
\code{pattern} or \link[quanteda]{dictionary} values}

\item{verbose}{logical - report the documents/features that had no overlap with the provided pre-trained embeddings}
}
\value{
a \linkS4class{conText} object
}
\description{
Estimate an embedding regression model
}
\examples{

## given a target word
model1 <- conText(formula = immigration ~ party, # equivalent to: immigration ~ . (since party is the only covariate in this corpus)
                  corpus = cr_corpus_sample,
                  pre_trained = glove_subset,
                  transform = TRUE, transform_matrix = khodakA,
                  bootstrap = TRUE, num_bootstraps = 10,
                  stratify = TRUE,
                  permute = TRUE, num_permutations = 100,
                  window = 6, valuetype = "fixed",
                  case_insensitive = TRUE,
                  verbose = TRUE)

# the party covariate is automatically "dummified"
rownames(model1)

# the beta coefficient in this case corresponds to the alc embedding
# of immigration for Republican party speeches
model1[,1:5]

# (normed) coefficient table
model1@normed_cofficients

## embedding full documents, not just a target word

model1 <- conText(formula = . ~ party,
                  corpus = cr_corpus_sample,
                  pre_trained = glove_subset,
                  transform = TRUE, transform_matrix = khodakA,
                  bootstrap = TRUE, num_bootstraps = 10,
                  stratify = TRUE,
                  permute = TRUE, num_permutations = 100,
                  window = 6, valuetype = "fixed",
                  case_insensitive = TRUE,
                  verbose = TRUE)

}
\keyword{conText}
