% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{context_log_start}
\alias{context_log_start}
\alias{context_log_stop}
\title{Start and stop log}
\usage{
context_log_start()

context_log_stop()
}
\description{
Start and stop the context log.  Soon this might swap out for
\code{loggr}, but for now this should do.  When active, some
actions will print diagnostic information to the message stream.
This is used in particular by the program installed by
\code{\link{install_context}}.
}
\details{
The interface here will change by adding arguments.  Future versions
may support logging to a file.
}

