% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{context_log}
\alias{context_log}
\title{Send entry to context log}
\usage{
context_log(topic, value)
}
\arguments{
\item{topic}{Up to 9 character text string with the log topic}

\item{value}{Character string with the log entry}
}
\description{
Send an entry to the context log.  This is designed primarily for
use with pacakges that build off of context, so that they can log
in a consistent way.
}
