% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_nns.R
\name{contrast_nns}
\alias{contrast_nns}
\title{Contrast nearest neighbors}
\usage{
contrast_nns(
  x,
  groups = NULL,
  pre_trained = NULL,
  transform = TRUE,
  transform_matrix = NULL,
  bootstrap = TRUE,
  num_bootstraps = 20,
  permute = TRUE,
  num_permutations = 100,
  candidates = NULL,
  N = 20,
  verbose = TRUE
)
}
\arguments{
\item{x}{(quanteda) tokens object}

\item{groups}{a variable with two unique values (can be numeric, factor, character)}

\item{pre_trained}{a F x D matrix of numeric values corresponding to pretrained embeddings
F = number of features and D = embedding dimensions.
rownames(pre_trained) = set of features for which there is a pre-trained embedding}

\item{transform}{(logical) if TRUE (default) apply the \verb{a la carte} transformation,
if FALSE ouput untransformed averaged embedding}

\item{transform_matrix}{a D x D matrix of numeric values corresponding
to the `a la carte`` transformation matrix}

\item{bootstrap}{logical - if TRUE, bootstrap cosine similarity ratio - required to get standard errors around cosine similarity ratios}

\item{num_bootstraps}{numeric - number of bootstraps to use}

\item{permute}{logical - if TRUE, compute empirical p-values using permutation test}

\item{num_permutations}{numeric - number of permutations to use}

\item{candidates}{character vector defining the candidates for nearest neighbors}

\item{N}{numeric (default is NULL). If defined,then nearest neighbors are subset to the union of the N neighbors of each group}

\item{verbose}{logical - report the documents/features that had
no overlap with the provided pre-trained embeddings}
}
\value{
a \code{data.frame} with following columns:
\item{\code{feature}}{(character) vector of features from the candidate set,
one instance for each target.}
\item{\code{value}}{(numeric) ratio of cosine similarities
(mean of boostraps if boostrap = TRUE).}
\item{\code{std.error}}{(numeric) sd/srt(num_bootstrap) of bootstrapped ratio
of cosine similarities if bootstrap = TRUE, if FALSE, column is dropped.}
\item{\code{p.value}}{(numeric) empirical p-value.}
}
\description{
Contrast nearest neighbors
}
\examples{

library(quanteda)

cr_toks <- tokens(cr_sample_corpus)

immig_toks <- tokens_context(x = cr_toks,
pattern = "immigration", window = 6L, hard_cut = FALSE, verbose = TRUE)

set.seed(42L)
party_nns <- contrast_nns(x = immig_toks,
groups = docvars(immig_toks, 'party'),
pre_trained = cr_glove_subset,
transform = TRUE, transform_matrix = cr_transform,
bootstrap = TRUE, num_bootstraps = 10,
permute = TRUE, num_permutations = 100,
candidates = NULL, N = 20,
verbose = FALSE)
}
\keyword{contrast_nns}
