% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_sim.R
\name{cos_sim}
\alias{cos_sim}
\title{Given a set of embeddings and a set of features, calculate cosine similarities.}
\usage{
cos_sim(x, pre_trained, features = NULL, as_list = TRUE)
}
\arguments{
\item{x}{a \linkS4class{dem} or \linkS4class{fem} object}

\item{pre_trained}{a F x D matrix of numeric values corresponding to pretrained embeddings
F = number of features and D = embedding dimensions.
rownames(pre_trained) = set of features for which there is a pre-trained embedding}

\item{features}{(character) features of interest}

\item{as_list}{(logical) if FALSE all results are combined into a single data.frame
If TRUE, a list of data.frames is returned with one data.frame per feature.}
}
\value{
a \code{data.frame} or list of data.frames (one for each feature)
with the following columns:
\item{\code{target}}{ (character) vector with the rownames of the dfm,
either defining the groups or the target terms}.
\item{\code{feature}}{(character) vector of feature terms, one
instance for each target.}
\item{\code{value}}{(numeric) cosine similarity between target
and feature.}
}
\description{
Given a set of embeddings and a set of features, calculate cosine similarities.
}
\examples{

library(quanteda)

# tokenize corpus
toks <- tokens(cr_sample_corpus)

# build a tokenized corpus of contexts sorrounding a target term
immig_toks <- tokens_context(x = toks, pattern = "immigr*", window = 6L)

# build document-feature matrix
immig_dfm <- dfm(immig_toks)

# construct document-embedding-matrix
immig_dem <- dem(immig_dfm, pre_trained = cr_glove_subset,
transform = TRUE, transform_matrix = cr_transform, verbose = FALSE)

# to get group-specific embeddings, average within party
immig_wv_party <- dem_group(immig_dem, groups = immig_dem@docvars$party)

# compute the cosine similarity between each party's embedding and a specific set of features
cos_sim(immig_wv_party, pre_trained = cr_glove_subset,
features = c('reform', 'enforcement'), as_list = FALSE)
}
\keyword{cos_sim}
