% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cr_sample_corpus}
\alias{cr_sample_corpus}
\title{Congressional Record sample corpus}
\format{
A quanteda corpus with 200 documents and 3 docvars:
\describe{
\item{party}{party of speaker, (D)emocrat or (R)epublican}
\item{gender}{gender of speaker, (F)emale or (M)ale}
\item{session_id}{id of Congress session in which speech was given}
...
}
}
\source{
\url{https://data.stanford.edu/congress_text}
}
\usage{
cr_sample_corpus
}
\description{
A (quanteda) corpus containing a sample of the United States Congressional Record
(daily transcripts) covering the 111th to 114th Congresses.
The raw corpus is first subset to speeches
containing the regular expression "immig*". Then 100 docs from each party-gender pair
is randomly sampled. For full data and pre-processing file, see:
https://www.dropbox.com/sh/jsyrag7opfo7l7i/AAB1z7tumLuKihGu2-FDmhmKa?dl=0
}
\keyword{datasets}
