% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dem-class}
\alias{dem-class}
\title{Virtual class "dem" for a document-embedding matrix}
\description{
The \verb{dem class} is \verb{dgCMatrix class} matrix with additional slots:
}
\section{Slots}{

\describe{
\item{\code{docvars}}{document covariates, inherited from dfm and corpus
subset to embeddable documents}

\item{\code{features}}{features used in computing the document embeddings}
}}

\seealso{
\code{dem}
}
\keyword{dem}
\keyword{internal}
