% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_group.R
\name{dem_group}
\alias{dem_group}
\title{Average document-embeddings in a dem by a grouping variable}
\usage{
dem_group(x, groups = NULL)
}
\arguments{
\item{x}{a \link{dem}}

\item{groups}{a character or factor variable equal in length to the number of documents}
}
\value{
a \linkS4class{dem} object
}
\description{
Average embeddings in a dem by a grouping variable, by averaging over columns within groups
and creating new "documents" with the group labels.
Similar in essence to \link{dfm_group}.
}
\examples{
corpus_dfm <- create_dfm(cr_corpus_sample, features_remove = quanteda::stopwords('en'))
corpus_dem <- dem(corpus_dfm, pre_trained = glove_subset, transform = TRUE, transform_matrix = khodakA, verbose = TRUE)
corpus_dem_grouped <- dem_group(x = corpus_dem, groups = corpus_dem@docvars$party)
}
\keyword{dem_group}
