% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_vocab.R
\name{get_local_vocab}
\alias{get_local_vocab}
\title{Get local vocab}
\usage{
get_local_vocab(context, pre_trained)
}
\arguments{
\item{context}{the text to be used to compute localized embeddings}

\item{pre_trained}{the pre-trained embeddings to be used to compute localized embeddings}
}
\value{
chr vector - vector of words in vocabulary
}
\description{
Local vocab consists of the intersect between the set of pre-trained embeddings
and the local contexts.
}
