% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{\%as\%}
\alias{\%as\%}
\title{Context manager's `as` function}
\usage{
context \%as\% variables
}
\arguments{
\item{context}{ContextManager S3 class or a character string.}

\item{variables}{Variable name to use within the context or a variable name
in which to store the `context` if it is not a ContextManager S3 class.}
}
\value{
ContextAs S3 object  with elements `context` and `variable`.
}
\description{
Context manager's `as` function
}
