% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_packages}
\alias{install_packages}
\title{Install packages}
\usage{
install_packages(packages, package_sources, ..., error = TRUE)
}
\arguments{
\item{packages}{A character vector of packages to install}

\item{package_sources}{Source information from
\code{\link{package_sources}}.}

\item{...}{Additional arguments passed through to
\code{\link{install.packages}}.}

\item{error}{Throw an error if package installation fails.}
}
\description{
Install packages, using source information from
\code{\link{package_sources}}.  In contrast with
\code{install.packages} this function can install packages in a
new local repository (used to isolate packages from the rest of
the system) and can install packages originally found on github
and bitbucket by creating a transient drat repository.  In
addition, \code{install_packages_missing} only installs
\emph{missing} packages only.  Also in contrast with
\code{install.packages}, these function will throw an error if
package installation fails (unless error=FALSE is set).
}
\section{Warning}{


Because of the possibility of confusion with
  \code{\link{install.packages}}, this function may be renamed
  shortly.
}

