% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nns_ratio.R
\name{nns_ratio}
\alias{nns_ratio}
\title{Computes the ratio of cosine similarities for two embeddings over
the union of their respective top N nearest neighbors.}
\usage{
nns_ratio(x, N = 10, candidates = character(0), pre_trained)
}
\arguments{
\item{x}{a \linkS4class{dem} or \linkS4class{fem} object or in general a matrix of embeddings}

\item{N}{number of nearest neighbors to consider
Note: if nearest neighbors overlap, the resulting number of features
will be fewer than 2*N.}

\item{candidates}{character vector delimiting the features to consider as nearest neighbor candidates}

\item{pre_trained}{a F x D matrix of numeric values corresponding to pretrained embeddings
F = number of features and D = embedding dimensions.
rownames(pre_trained) = set of features for which there is a pre-trained embedding}
}
\value{
a \code{data.frame} with following columns:
\item{\code{feature}}{(character) vector of features from the candidate set,
one instance for each target.}
\item{\code{value}}{(numeric) ratio of cosine similarities}.
}
\description{
Computes the ratio of cosine similarities for two embeddings over
the union of their respective top N nearest neighbors.
}
\examples{

library(quanteda)

# build corpus of contexts around immigration
immig_corpus <- corpus_context(x = cr_sample_corpus,
pattern = "immigration",
window = 6L,
verbose = TRUE)

# tokenize text
immig_toks <- tokens(immig_corpus)

# construct document-feature-matrix
immig_dfm <- dfm(immig_toks)

# construct document-embedding-matrix
immig_dem <- dem(immig_dfm,
pre_trained =
glove_subset,
transform = TRUE,
transform_matrix = khodakA,
verbose = FALSE)

# group document-embedding-matrix
immig_dem_party <- dem_group(immig_dem, groups = immig_dem@docvars$party)

# find nearest neighbors
nns_ratio(x = immig_dem_party, pre_trained = glove_subset, candidates = character(0), N = 10)
}
\keyword{nns_ratio}
