% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{open}
\alias{open}
\title{Open file implementation using a context manager}
\usage{
open(file_name, mode = "r")
}
\arguments{
\item{file_name}{File name to open.}

\item{mode}{File mode.}
}
\value{
S3 class ContextManager
}
\description{
Similar to Python's `open()`, intended to be used with a context manager.
}
