% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nns_ratio.R
\name{plot_nns_ratio}
\alias{plot_nns_ratio}
\title{Plot output of get_nns_ratio}
\usage{
plot_nns_ratio(x, alpha = 0.01, horizontal = TRUE)
}
\arguments{
\item{x}{output of get_nns_ratio}

\item{alpha}{(numerical) betwee 0 and 1;significance threshold to identify significant values}

\item{horizontal}{(logical) defines the type of plot}
}
\value{
a ggplot
}
\description{
Plot output of get_nns_ratio
}
\examples{

library(ggplot2)
library(quanteda)

# build corpus of contexts around immigration
immig_corpus <- corpus_context(x = cr_sample_corpus,
pattern = "immigration",
window = 6L,
verbose = TRUE)

local_vocab <- get_local_vocab(as.character(immig_corpus), pre_trained = glove_subset)

set.seed(42L)
nns_ratio_immig <- get_nns_ratio(x = immig_corpus,
N = 20,
groups = docvars(immig_corpus, 'party'),
numerator = "R",
candidates = local_vocab,
pre_trained = glove_subset,
transform = TRUE,
transform_matrix = khodakA,
bootstrap = TRUE,
num_bootstraps = 10,
permute = TRUE,
num_permutations = 10)

plot_nns_ratio(x = nns_ratio_immig, alpha = 0.01, horizontal = TRUE)
}
\keyword{plot_nns_ratio}
