% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nns_ratio.R
\name{plot_nns_ratio}
\alias{plot_nns_ratio}
\title{Plot output of get_nns_ratio}
\usage{
plot_nns_ratio(x, alpha = 0.01, horizontal = TRUE)
}
\arguments{
\item{x}{output of get_nns_ratio}

\item{alpha}{(numerical) betwee 0 and 1;significance threshold to identify significant values}

\item{horizontal}{(logical) defines the type of plot}
}
\value{
a ggplot
}
\description{
Plot output of get_nns_ratio
}
\examples{

library(ggplot2)
library(quanteda)

# tokenize corpus
cr_toks <- tokens(cr_sample_corpus)

# get tokens around immigration
immig_toks <- tokens_context(x = cr_toks,
pattern = "immigration", window = 6L, hard_cut = FALSE, verbose = TRUE)

set.seed(42L)
nns_ratio_immig <- get_nns_ratio(x = immig_toks,
N = 20,
groups = docvars(immig_toks, 'party'),
numerator = "R",
candidates = character(0),
pre_trained = cr_glove_subset,
transform = TRUE,
transform_matrix = cr_transform,
bootstrap = TRUE,
num_bootstraps = 10,
permute = TRUE,
num_permutations = 10,
verbose = TRUE)

plot_nns_ratio(x = nns_ratio_immig, alpha = 0.01, horizontal = TRUE)
}
\keyword{plot_nns_ratio}
