% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototypical_context.R
\name{prototypical_context}
\alias{prototypical_context}
\title{Find prototypical contexts (most similar on average to the full set of contexts)}
\usage{
prototypical_context(
  context,
  pre_trained,
  transform_matrix,
  N = 3,
  norm = "l2"
)
}
\arguments{
\item{context}{vector of texts - `context`` variable in get_context output}

\item{pre_trained}{a V x D matrix of numeric values - pretrained embeddings with V = size of vocabulary and D = embedding dimensions}

\item{transform_matrix}{a D x D transformation matrix}

\item{N}{number of contexts to return}

\item{norm}{character = c("l2", "none") - set to 'l2' for cosine similarity and to 'none' for inner product (see ?sim2 in text2vec)}
}
\value{
character vector of contexts ordered by average similarity to all contexts
}
\description{
Find prototypical contexts (most similar on average to the full set of contexts)
}
