% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provision.R
\name{provision_context}
\alias{provision_context}
\title{Provision a context}
\usage{
provision_context(ctx, platform = NULL, version = NULL, quiet = FALSE,
  allow_missing = FALSE, installed_action = "skip",
  additional_libraries = NULL)
}
\arguments{
\item{ctx}{A context object.  We'll look in here for packages to install.}

\item{platform}{Platform to create the library for}

\item{version}{Version of R to create the library for}

\item{quiet}{Be quiet when installing source packages}

\item{allow_missing}{Allow packages requiring compilation to be
skipped over?  If \code{TRUE} you'll need to sort these out
yourself.}

\item{installed_action}{Action if packages are installed}
}
\description{
Provision a context
}

