% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-by-predicate.r
\name{scale_by_predicate}
\alias{scale_by_predicate}
\title{Expand variables in a data frame via conceptual scaling}
\usage{
scale_by_predicate(data, predicate)
}
\arguments{
\item{data}{Data frame.}

\item{predicate}{A logical expression involving variables in \code{data}.}
}
\value{
A binary matrix (formal context).
}
\description{
Expand variables in a data frame via conceptual scaling
}

