% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{save_task}
\alias{load_task}
\alias{save_task}
\title{Sve and reload tasks}
\usage{
save_task(expr, context = NULL, envir = parent.frame(), root = tempdir())

load_task(handle, envir = .GlobalEnv)
}
\arguments{
\item{expr}{An expression to save}

\item{context}{Either a \code{context} or a \code{context_handle}
object.  If missing (or NULL) one will be automatically
generated, using \code{envir} and \code{root}.  See
\code{\link{save_context}}.}

\item{envir}{Passed through to \code{save_context} when creating
contexts automatically.}

\item{root}{Root directory to store and retrieve files.}

\item{handle}{A handle to load the task}
}
\description{
Save and reload tasks.  Tasks consist of an expression bound to a
\code{context}.
}

