% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_save}
\alias{task_save}
\title{Save and reload tasks}
\usage{
task_save(expr, context, envir = parent.frame())
}
\arguments{
\item{expr}{An expression to save}

\item{context}{A \code{context} object}

\item{envir}{Passed through to \code{context_save} when locating
local variables.}
}
\value{
An identifier that can be used to retrieve or run the task
  later.  This is simply a short string.
}
\description{
Save and reload tasks.  Tasks consist of an expression bound to a
\code{context}.
}

