% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{save_task}
\alias{load_task}
\alias{read_task}
\alias{save_task}
\alias{save_task_list}
\title{Save and reload tasks}
\usage{
save_task(expr, context = NULL, envir = parent.frame(), root = tempdir())

save_task_list(list, context = NULL, envir = parent.frame(),
  root = tempdir())

load_task(handle, install = TRUE, envir = .GlobalEnv)

read_task(handle)
}
\arguments{
\item{expr}{An expression to save}

\item{context}{Either a \code{context} or a \code{context_handle}
object.  If missing (or NULL) one will be automatically
generated, using \code{envir} and \code{root}.  See
\code{\link{save_context}}.}

\item{envir}{Passed through to \code{save_context} when creating
contexts automatically.}

\item{root}{Root directory to store and retrieve files.}

\item{list}{a \code{list} of tasks, all to be evaluted in the same context.}

\item{handle}{A handle to load the task}

\item{install}{Should missing packages be installed?}
}
\description{
Save and reload tasks.  Tasks consist of an expression bound to a
\code{context}.
}

