% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_save}
\alias{task_load}
\alias{task_read}
\alias{task_save}
\alias{task_save_list}
\alias{tasks_list}
\title{Save and reload tasks}
\usage{
task_save(expr, context, envir = parent.frame())

task_save_list(list, context, envir = parent.frame())

task_load(handle, install = TRUE, envir = .GlobalEnv, load_context = TRUE)

task_read(handle)

tasks_list(root)
}
\arguments{
\item{expr}{An expression to save}

\item{context}{Either a \code{context} or a \code{context_handle}
object.  If missing (or NULL) one will be automatically
generated, using \code{envir} and \code{root}.  See
\code{\link{context_save}}.}

\item{envir}{Passed through to \code{context_save} when locating
local variables.}

\item{list}{a \code{list} of tasks, all to be evaluted in the same context.}

\item{handle}{A handle to load the task}

\item{install}{Should missing packages be installed?}

\item{load_context}{Load the context?  This is generally what is
wanted, but if you're repeatedly evaluating expressions in the
same context.}

\item{root}{root}
}
\description{
Save and reload tasks.  Tasks consist of an expression bound to a
\code{context}.
}

