% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_delete}
\alias{task_delete}
\title{Delete a task}
\usage{
task_delete(ids, root)
}
\arguments{
\item{ids}{Vector of task ids}

\item{root}{A context root (not just the db as in
\code{\link{task_result}} as we need to know the actual path to
the root).  A \code{context} object is also OK.}
}
\value{
\code{TRUE} if a task was actually deleted.
}
\description{
Delete a task, including its results.
}

