% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_query.R
\name{task_expr}
\alias{task_expr}
\title{Fetch task expression}
\usage{
task_expr(id, db, locals = FALSE)
}
\arguments{
\item{id}{Single task identifier}

\item{db}{Something that can be converted to a context db object
(a database, root or context).}

\item{locals}{Return locals bound to the expression (as an
attribute "locals")}
}
\description{
Fetch expression for a task
}

