% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_query.R
\name{task_expr}
\alias{task_expr}
\alias{task_function_name}
\title{Fetch task expression}
\usage{
task_expr(id, root, locals = FALSE)

task_function_name(id, root)
}
\arguments{
\item{locals}{Return locals bound to the expression (as an
attribute "locals")}

\item{handle}{A task handle}
}
\description{
Fetch expression for a task
}

