% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_query.R
\name{task_log}
\alias{task_log}
\title{Return task log}
\usage{
task_log(id, root, parse = TRUE)
}
\arguments{
\item{id}{Single task identifier}

\item{root}{A context root (not just the db as in
\code{\link{task_result}} as we need to know the actual path to
the root).  A \code{context} object is also OK.}
}
\description{
Return the log of a task, if enabled.
}
\details{
The returned object is of class \code{task_log}, which has a print
method that will nicely display.  Output is grouped into phases.
}

