% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_query.R
\name{task_log}
\alias{task_log}
\title{Return task log}
\usage{
task_log(id, root)
}
\description{
Return the log of a task, if enabled.
}
\details{
The returned object is of class \code{task_log}, which has a print
method that will nicely display.  Output is grouped into phases.
}

