% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_query.R
\name{task_result}
\alias{task_result}
\title{Fetch task result}
\usage{
task_result(id, root, sanitise = FALSE)
}
\arguments{
\item{sanitise}{Should we avoid throwing an error if a task is not
completed?  Used internally, and not generally needed.}

\item{handle}{A task handle}
}
\description{
Fetch result from completed task.
}

