% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_status}
\alias{task_status}
\title{Task status}
\usage{
task_status(handle, named = FALSE)
}
\arguments{
\item{handle}{Task handle}

\item{named}{Name the output with the task ids?}
}
\description{
Task status
}

