% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{write_context_script}
\alias{write_context_script}
\alias{parse_context_args}
\title{Write context-using script}
\usage{
write_context_script(path, versioned = TRUE)

parse_context_args(args, name, nargs)
}
\arguments{
\item{path}{Path to the root of the context}

\item{args}{Arguments passed to the script, as returned by
\code{commandArgs(TRUE)}.  Will be a character vector}

\item{name}{Name of the script; this will be the filename}

\item{nargs}{The number of arguments, in addition to the context
root, that the script takes.  This can be a vector of length two
where minimum and maximum numbers of arguments are supported.}

\item{target}{The name function to call.  This must be at least
namespace qualified ("::") but probably will be a hidden
function (":::")}
}
\description{
Write a very simple context script.  This can be used for
applications that need to use context, including setting up the
library the context is itself found in.  No fancy argument
matching will be done (e.g., optparse, docopt).
}
